Start cleaning up changes of origxyz
