+++ utils/mrisurf_base.c
@@ -20,6 +20,25 @@
 #include "mrisurf_base.h"
 
 
+int mris_sort_compare_float(const void *pc1, const void *pc2)
+{
+  register float c1, c2;
+
+  c1 = *(float *)pc1;
+  c2 = *(float *)pc2;
+
+  /*  return(c1 > c2 ? 1 : c1 == c2 ? 0 : -1) ;*/
+  if (c1 > c2) {
+    return (1);
+  }
+  else if (c1 < c2) {
+    return (-1);
+  }
+
+  return (0);
+}
+
+
 int  MRIS_acquireTemp(MRIS* mris, MRIS_TempAssigned temp) {
   int const bits = 1 << temp;
   int const * tc = &mris->tempsAssigned;
