+++ utils/mrisurf_metricProperties.c
@@ -1,8 +1,5 @@
+#define COMPILING_MRISURF_METRIC_PROPERTIES
 #define COMPILING_MRISURF_TOPOLOGY_FRIEND_CHECKED
-/*
- * @file utilities operating on Original
- *
- */
 /*
  * surfaces Author: Bruce Fischl, extracted from mrisurf.c by Bevin Brett
  *
@@ -68,7 +65,7 @@ void INTEGRATION_PARMS_copyFp (INTEGRATION_PARMS* dst, INTEGRATION_PARMS const *
 int mrisurf_activeRealmTreesSize;
 int mrisurf_orig_clock;
 
-int MRISprintVertexStats(MRI_SURFACE *mris, int vno, FILE *fp, int which_vertices)
+int MRISprintVertexStats(MRIS *mris, int vno, FILE *fp, int which_vertices)
 {
   double mn, d;
   int n, num;
@@ -107,7 +104,18 @@ int MRISprintVertexStats(MRI_SURFACE *mris, int vno, FILE *fp, int which_vertice
 }
 
 
-int load_orig_triangle_vertices(MRI_SURFACE *mris, int fno, double U0[3], double U1[3], double U2[3])
+void MRISsetOriginalXYZ(MRIS *mris, int vno, float origx, float origy, float origz) 
+{
+  cheapAssertValidVno(mris,vno);
+  VERTEX * v = mris->vertices[vno];
+    
+  const float * pcx = &v-=>origx;  float * px = (float*)pcx; *pxm = origx;
+  const float * pcy = &v-=>origy;  float * py = (float*)pcy; *pym = origy;
+  const float * pcz = &v-=>origz;  float * pz = (float*)pcz; *pzm = origz;
+}
+
+
+int load_orig_triangle_vertices(MRIS *mris, int fno, double U0[3], double U1[3], double U2[3])
 {
   VERTEX *v;
   FACE *face;
@@ -128,7 +136,7 @@ int load_orig_triangle_vertices(MRI_SURFACE *mris, int fno, double U0[3], double
   return (NO_ERROR);
 }
 
-int load_triangle_vertices(MRI_SURFACE *mris, int fno, double U0[3], double U1[3], double U2[3], int which)
+int load_triangle_vertices(MRIS *mris, int fno, double U0[3], double U1[3], double U2[3], int which)
 {
   VERTEX *v;
   FACE *face;
@@ -183,7 +191,7 @@ int load_triangle_vertices(MRI_SURFACE *mris, int fno, double U0[3], double U1[3
 }
 
 
-int MRISextractVertexCoords(MRI_SURFACE *mris, float *locations[3], int which)
+int MRISextractVertexCoords(MRIS *mris, float *locations[3], int which)
 {
   int vno, nvertices;
   VERTEX *v;
@@ -255,7 +263,7 @@ int MRISextractVertexCoords(MRI_SURFACE *mris, float *locations[3], int which)
   return (NO_ERROR);
 }
 
-int MRISimportVertexCoords(MRI_SURFACE *mris, float *locations[3], int which)
+int MRISimportVertexCoords(MRIS *mris, float *locations[3], int which)
 {
   int vno, nvertices;
   VERTEX *v;
@@ -330,11 +338,11 @@ int MRISimportVertexCoords(MRI_SURFACE *mris, float *locations[3], int which)
 
 /*-----------------------------------------------------*/
 /*!
-  \fn int MRISreverseCoords(MRI_SURFACE *mris, int which_reverse, int reverse_face_order, int which_coords)
+  \fn int MRISreverseCoords(MRIS *mris, int which_reverse, int reverse_face_order, int which_coords)
   \brief Reverse sign of one of the dimensions of the surface coords.
   If reversing X, the order of the verticies is also reversed.
 */
-int MRISreverseCoords(MRI_SURFACE *mris, int which_direction, int reverse_face_order, int which_coords)
+int MRISreverseCoords(MRIS *mris, int which_direction, int reverse_face_order, int which_coords)
 {
   int vno;
   float x = 0, y = 0, z = 0;
@@ -406,11 +414,11 @@ int MRISreverseCoords(MRI_SURFACE *mris, int which_direction, int reverse_face_o
 }
 /*-----------------------------------------------------*/
 /*!
-  \fn int MRISreverse(MRI_SURFACE *mris, int which)
+  \fn int MRISreverse(MRIS *mris, int which)
   \brief Reverse sign of one of the dimensions of the surface coords.
   If reversing X, the order of the verticies is also reversed.
 */
-int MRISreverse(MRI_SURFACE *mris, int which, int reverse_face_order)
+int MRISreverse(MRIS *mris, int which, int reverse_face_order)
 {
   int vno;
   float x, y, z;
@@ -448,7 +456,7 @@ int MRISreverse(MRI_SURFACE *mris, int which, int reverse_face_order)
   return (NO_ERROR);
 }
 
-int MRISscaleVertexCoordinates(MRI_SURFACE *mris, double scale)
+int MRISscaleVertexCoordinates(MRIS *mris, double scale)
 {
   VERTEX *v;
   int vno;
@@ -477,6 +485,1954 @@ int MRISscale(MRIS *mris, double scale)
     v->z *= scale;
   }
   return (0);
+}
+
+int mrisFlipPatch(MRIS *mris)
+{
+  int vno;
+  VERTEX *v;
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->x = -v->x;
+  }
+
+  return (NO_ERROR);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISaverageVertexPositions(MRIS *mris, int navgs)
+{
+  int i, vno, vnb, vnum;
+  float x, y, z, num;
+
+  for (i = 0; i < navgs; i++) {
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+      VERTEX                * const v  = &mris->vertices         [vno];
+      if (v->ripflag) {
+        continue;
+      }
+      x = v->x;
+      y = v->y;
+      z = v->z;
+      int const * pnb  = vt->v;
+      vnum = vt->vnum;
+      for (num = 0.0f, vnb = 0; vnb < vnum; vnb++) {
+        VERTEX * const vn = &mris->vertices[*pnb++]; /* neighboring vertex pointer */
+        if (vn->ripflag) {
+          continue;
+        }
+        num++;
+        x += vn->x;
+        y += vn->y;
+        z += vn->z;
+      }
+      num++; /* account for central vertex */
+      v->tdx = x / num;
+      v->tdy = y / num;
+      v->tdz = z / num;
+    }
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      VERTEX * const v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      v->x = v->tdx;
+      v->y = v->tdy;
+      v->z = v->tdz;
+    }
+  }
+  return (NO_ERROR);
+}
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISaverageEveryOtherVertexPositions(MRIS *mris, int navgs, int which)
+{
+  int i, vno, vnb, vnum;
+  float x, y, z, num;
+
+  which = ISODD(which);
+  for (i = 0; i < navgs; i++) {
+    for (vno = which; vno < mris->nvertices; vno += 2) {
+      VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+      VERTEX                * const v  = &mris->vertices         [vno];
+      if (v->ripflag) {
+        continue;
+      }
+      x = v->x;
+      y = v->y;
+      z = v->z;
+      int const * pnb  = vt->v;
+      vnum = vt->vnum;
+      for (num = 0.0f, vnb = 0; vnb < vnum; vnb++) {
+        VERTEX const * const vn = &mris->vertices[*pnb++]; /* neighboring vertex pointer */
+        if (vn->ripflag) {
+          continue;
+        }
+        num++;
+        x += vn->x;
+        y += vn->y;
+        z += vn->z;
+      }
+      num++; /* account for central vertex */
+      v->tdx = x / num;
+      v->tdy = y / num;
+      v->tdz = z / num;
+    }
+    for (vno = which; vno < mris->nvertices; vno += 2) {
+      VERTEX * const v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      v->x = v->tdx;
+      v->y = v->tdy;
+      v->z = v->tdz;
+    }
+  }
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISsequentialAverageVertexPositions(MRIS *mris, int navgs)
+{
+  int i, vno, vnb, vnum;
+  float x, y, z, num;
+
+  for (i = 0; i < navgs; i++) {
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+      VERTEX                * const v  = &mris->vertices         [vno];
+      if (v->ripflag || v->marked) {
+        continue;
+      }
+      x = v->x;
+      y = v->y;
+      z = v->z;
+      int const * pnb  = vt->v;
+      vnum = vt->vnum;
+      for (num = 0.0f, vnb = 0; vnb < vnum; vnb++) {
+        VERTEX const * const vn = &mris->vertices[*pnb++]; /* neighboring vertex pointer */
+        if (vn->ripflag) {
+          continue;
+        }
+        num++;
+        x += vn->x;
+        y += vn->y;
+        z += vn->z;
+      }
+      num++; /* account for central vertex */
+      v->x = x / num;
+      v->y = y / num;
+      v->z = z / num;
+    }
+  }
+  return (NO_ERROR);
+}
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISscaleDistances(MRIS *mris, float scale)
+{
+  int vno, n;
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+    VERTEX                * const v  = &mris->vertices         [vno];
+    if (v->ripflag) {
+      continue;
+    }
+    for (n = 0; n < vt->vtotal; n++) {
+      v->dist[n] *= scale;
+    }
+  }
+  return (NO_ERROR);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+MRIS *MRISscaleBrain(MRIS *mris_src, MRIS *mris_dst, float scale)
+{
+  VERTEX *v;
+  int k;
+  float xlo, ylo, zlo, xhi, yhi, zhi, x0, y0, z0;
+
+  /*  mris_dst = MRIScenter(mris_src, mris_dst) ;*/
+  xhi = yhi = zhi = -10000;
+  xlo = ylo = zlo = 10000;
+
+  /* find the center */
+  for (k = 0; k < mris_dst->nvertices; k++) {
+    v = &mris_dst->vertices[k];
+    if (v->ripflag) {
+//      continue;
+    }
+    if (v->x > xhi) {
+      xhi = v->x;
+    }
+    if (v->x < xlo) {
+      xlo = v->x;
+    }
+    if (v->y > yhi) {
+      yhi = v->y;
+    }
+    if (v->y < ylo) {
+      ylo = v->y;
+    }
+    if (v->z > zhi) {
+      zhi = v->z;
+    }
+    if (v->z < zlo) {
+      zlo = v->z;
+    }
+  }
+
+  /* scale around the center */
+  x0 = (xlo + xhi) / 2.0f;
+  y0 = (ylo + yhi) / 2.0f;
+  z0 = (zlo + zhi) / 2.0f;
+  for (k = 0; k < mris_dst->nvertices; k++) {
+    v = &mris_dst->vertices[k];
+    v->x = (v->x - x0) * scale + x0;
+    v->y = (v->y - y0) * scale + y0;
+    v->z = (v->z - z0) * scale + z0;
+    if (v->x > xhi) {
+      xhi = v->x;
+    }
+    if (v->x < xlo) {
+      xlo = v->x;
+    }
+    if (v->y > yhi) {
+      yhi = v->y;
+    }
+    if (v->y < ylo) {
+      ylo = v->y;
+    }
+    if (v->z > zhi) {
+      zhi = v->z;
+    }
+    if (v->z < zlo) {
+      zlo = v->z;
+    }
+  }
+
+  /* recompute the dimensions */
+  xhi = yhi = zhi = -10000;
+  xlo = ylo = zlo = 10000;
+  for (k = 0; k < mris_dst->nvertices; k++) {
+    v = &mris_dst->vertices[k];
+    if (v->x > xhi) {
+      xhi = v->x;
+    }
+    if (v->x < xlo) {
+      xlo = v->x;
+    }
+    if (v->y > yhi) {
+      yhi = v->y;
+    }
+    if (v->y < ylo) {
+      ylo = v->y;
+    }
+    if (v->z > zhi) {
+      zhi = v->z;
+    }
+    if (v->z < zlo) {
+      zlo = v->z;
+    }
+  }
+  mris_dst->xlo = xlo;
+  mris_dst->ylo = ylo;
+  mris_dst->zlo = zlo;
+  mris_dst->xctr = (xhi + xlo) / 2;
+  mris_dst->yctr = (yhi + ylo) / 2;
+  mris_dst->zctr = (zhi + zlo) / 2;
+
+  return (mris_dst);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISscaleBrainArea(MRIS *mris)
+{
+  float scale;
+
+  scale = sqrt(mris->orig_area / (mris->total_area + mris->neg_area));
+  MRISscaleBrain(mris, mris, scale);
+  MRIScomputeMetricProperties(mris);
+  return (NO_ERROR);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  Translate a surface by (dx, dy, dz)
+  ------------------------------------------------------*/
+int MRIStranslate(MRIS *mris, float dx, float dy, float dz)
+{
+  int vno;
+  VERTEX *v;
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->x += dx;
+    v->y += dy;
+    v->z += dz;
+  }
+  mrisComputeSurfaceDimensions(mris);
+  return (NO_ERROR);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+MRIS *MRISrotate(MRIS *mris_src, MRIS *mris_dst, float alpha, float beta, float gamma)
+{
+  if (!mris_dst) {
+    mris_dst = MRISclone(mris_src);
+  }
+
+  float const sa = sin(alpha);
+  float const sb = sin(beta);
+  float const sg = sin(gamma);
+  float const ca = cos(alpha);
+  float const cb = cos(beta);
+  float const cg = cos(gamma);
+
+  float const cacb = ca * cb;
+  float const cacgsb = ca * cg * sb;
+  float const sasg = sa * sg;
+  float const cgsa = cg * sa;
+  float const casbsg = ca * sb * sg;
+  float const cbsa = cb * sa;
+  float const cgsasb = cg * sa * sb;
+  float const casg = ca * sg;
+  float const cacg = ca * cg;
+  float const sasbsg = sa * sb * sg;
+  float const cbcg = cb * cg;
+  float const cbsg = cb * sg;
+
+  int vno;
+  
+  ROMP_PF_begin
+#ifdef HAVE_OPENMP
+  #pragma omp parallel for if_ROMP(assume_reproducible) schedule(guided)
+#endif
+  for (vno = 0; vno < mris_src->nvertices; vno++) {
+    ROMP_PFLB_begin
+    
+    if (vno == Gdiag_no) {
+      DiagBreak();
+    }
+
+    VERTEX *vertex = &mris_src->vertices[vno];
+    float x = vertex->x;
+    float y = vertex->y;
+    float z = vertex->z;
+    float xp = x * cacb + z * (-cacgsb - sasg) + y * (cgsa - casbsg);
+    float yp = -x * cbsa + z * (cgsasb - casg) + y * (cacg + sasbsg);
+    float zp = z * cbcg + x * sb + y * cbsg;
+    vertex->x = xp;
+    vertex->y = yp;
+    vertex->z = zp;
+
+    ROMP_PFLB_end
+  }
+  ROMP_PF_end
+  
+  return (mris_dst);
+}
+
+
+/*------------------------------------------------------------------------
+  MRISmatrixMultiply() - simply multiplies matrix M by the vertex xyz.
+  See also MRIStransform().
+  ------------------------------------------------------------------------*/
+int MRISmatrixMultiply(MRIS *mris, MATRIX *M)
+{
+  int vno;
+
+  ROMP_PF_begin
+#ifdef HAVE_OPENMP
+  #pragma omp parallel for if_ROMP(experimental)
+#endif
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    ROMP_PFLB_begin
+    
+    VERTEX *v;
+    MATRIX *xyz, *Mxyz;
+    xyz = MatrixAlloc(4, 1, MATRIX_REAL);
+    xyz->rptr[4][1] = 1.0;
+    Mxyz = MatrixAlloc(4, 1, MATRIX_REAL);
+    v = &mris->vertices[vno];
+    xyz->rptr[1][1] = v->x;
+    xyz->rptr[2][1] = v->y;
+    xyz->rptr[3][1] = v->z;
+    MatrixMultiply(M, xyz, Mxyz);
+
+    v->x = Mxyz->rptr[1][1];
+    v->y = Mxyz->rptr[2][1];
+    v->z = Mxyz->rptr[3][1];
+    MatrixFree(&xyz);
+    MatrixFree(&Mxyz);
+    
+    ROMP_PFLB_end
+  }
+  ROMP_PF_end
+  
+  return (0);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  Scale a surface anisotropically.
+  ------------------------------------------------------*/
+int MRISanisotropicScale(MRIS *mris, float sx, float sy, float sz)
+{
+  VERTEX *v;
+  int k;
+  float x0, y0, z0;
+
+  mrisComputeSurfaceDimensions(mris);
+  /* scale around the center */
+  x0 = mris->xctr;
+  y0 = mris->yctr;
+  z0 = mris->zctr;
+
+  for (k = 0; k < mris->nvertices; k++) {
+    v = &mris->vertices[k];
+    if (v->ripflag) {
+      continue;
+    }
+    v->x = (v->x - x0) * sx + x0;
+    v->y = (v->y - y0) * sy + y0;
+    v->z = (v->z - z0) * sz + z0;
+  }
+
+  mrisComputeSurfaceDimensions(mris);
+  return (NO_ERROR);
+}
+
+
+/*-------------------------------------------------------*/
+/* Replaces x,y,z with theta,phi,radius. Assumes
+    that the surface xyz are already on the sphere.
+    Note: this is not related to the surface-based
+    spherical coords.
+ */
+int MRISsphericalCoords(MRIS *surf)
+{
+  int k;
+  double x, y, z, d2, d, r, theta, phi;
+
+  for (k = 0; k < surf->nvertices; k++) {
+    x = surf->vertices[k].x;
+    y = surf->vertices[k].y;
+    z = surf->vertices[k].z;
+    d2 = x * x + y * y;
+    d = sqrt(d2);
+    r = sqrt(d2 + z * z);
+    theta = atan2(y, x);
+    phi = atan2(z, d);
+    surf->vertices[k].x = theta;
+    surf->vertices[k].y = phi;
+    surf->vertices[k].z = r;
+  }
+  return (0);
+}
+
+
+// translate the COG of a surface to (0,0,0)
+void MRIScenterCOG2(MRIS *mris, double *xCOG, double *yCOG, double *zCOG)
+{
+  int k;
+  double x, y, z;
+  x = 0;
+  y = 0;
+  z = 0;
+  for (k = 0; k < mris->nvertices; k++) {
+    x += mris->vertices[k].x;
+    y += mris->vertices[k].y;
+    z += mris->vertices[k].z;
+  }
+  x /= mris->nvertices;
+  y /= mris->nvertices;
+  z /= mris->nvertices;
+  for (k = 0; k < mris->nvertices; k++) {
+    mris->vertices[k].x -= x;
+    mris->vertices[k].y -= y;
+    mris->vertices[k].z -= z;
+  }
+  if (xCOG && yCOG && zCOG) {
+    (*xCOG) = x;
+    (*yCOG) = y;
+    (*zCOG) = z;
+  }
+  /*       fprintf(stderr,"\nCOG Centered at x=%f y=%f z=%f",
+     (float)x,(float)y,(float)z);*/
+}
+
+void MRIScenterCOG(MRIS *mris) { MRIScenterCOG2(mris, NULL, NULL, NULL); }
+
+
+
+// smooth a surface 'niter' times with a step (should be around 0.5)
+void MRISsmoothSurface(MRIS *mris, int niter, float step)
+{
+  if (step > 1) step = 1.0f;
+
+  int iter;
+  for (iter = 0; iter < niter; iter++) {
+    MRIScomputeMetricProperties(mris);
+
+    int k;
+    for (k = 0; k < mris->nvertices; k++) {
+      VERTEX * v = &mris->vertices[k];
+      v->tx = v->x;
+      v->ty = v->y;
+      v->tz = v->z;
+    }
+
+    for (k = 0; k < mris->nvertices; k++) {
+      VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[k];
+      VERTEX                * const v  = &mris->vertices         [k];
+      int n = 0;
+      float x = 0, y = 0, z = 0;
+      int m;
+      for (m = 0; m < vt->vnum; m++) {
+        x += mris->vertices[vt->v[m]].tx;
+        y += mris->vertices[vt->v[m]].ty;
+        z += mris->vertices[vt->v[m]].tz;
+        n++;
+      }
+      x /= n;
+      y /= n;
+      z /= n;
+
+      v->x = v->x + step * (x - v->x);
+      v->y = v->y + step * (y - v->y);
+      v->z = v->z + step * (z - v->z);
+    }
+  }
+}
+
+
+/*-----------------------------------------------------
+  MRIStransform
+
+  Parameters: mris, mri, lta, mri_dst
+
+  Returns value: int
+
+  Description
+  Apply a linear transform (possibly octree) to a surface.
+  Note that the LT is given in MRI coordinates, so the
+  surface must be transformed into that coordinate system
+  before applying the linear transform
+  See also MRISmatrixMultiply().
+  ------------------------------------------------------*/
+#include "gcamorph.h"
+int MRIStransform(MRIS *mris, MRI *mri, TRANSFORM *transform, MRI *mri_dst)
+{
+  LTA *lta;
+  int vno;
+  VERTEX *v;
+  double xw, yw, zw;
+  MATRIX *m = 0;
+
+  // for ras-to-ras transform
+  MATRIX *RASFromSurfaceRAS = 0;
+  MATRIX *surfaceRASFromRAS = 0;
+  // for voxel-to-voxel transform
+  MATRIX *voxelFromSurfaceRAS = 0;
+  MATRIX *surfaceRASFromVoxel = 0;
+  //
+  MATRIX *surfaceRASFromSurfaceRAS = 0;
+  LT *lt = 0;
+  int srcPresent = 1;
+  int dstPresent = 1;
+  int error = NO_ERROR;
+  char errMsg[256];
+  int dstNotGiven = 0;
+  int srcNotGiven = 0;
+
+  if (transform->type == MORPH_3D_TYPE) {
+    GCA_MORPH *gcam;
+    double xs, ys, zs, xv, yv, zv;
+    float xv2, yv2, zv2;
+    VECTOR *v1, *v2;
+
+    /*
+      transform point from surface (ras) coords to the source (atlas) volume
+      that was used for the gcam:
+    */
+
+    gcam = (GCA_MORPH *)(transform->xform);
+    if (!mri_dst) {
+      dstNotGiven = 1;
+      mri_dst = MRIalloc(gcam->image.width, gcam->image.height, gcam->image.depth, MRI_UCHAR);
+      useVolGeomToMRI(&gcam->image, mri_dst);
+    }
+    if (!mri) {
+      srcNotGiven = 1;
+      mri = MRIalloc(gcam->atlas.width, gcam->atlas.height, gcam->atlas.depth, MRI_UCHAR);
+      useVolGeomToMRI(&gcam->atlas, mri);
+    }
+    if (gcam->type == GCAM_RAS) {
+        GCAMrasToVox(gcam, mri_dst);
+    }
+
+    v1 = VectorAlloc(4, MATRIX_REAL);
+    VECTOR_ELT(v1, 4) = 1.0;
+    v2 = VectorAlloc(4, MATRIX_REAL);
+    VECTOR_ELT(v2, 4) = 1.0;
+    voxelFromSurfaceRAS = voxelFromSurfaceRAS_(mri);
+
+    // now apply the transform
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+
+      // transform vertex point to actual voxel point
+      V3_X(v1) = v->x;
+      V3_Y(v1) = v->y;
+      V3_Z(v1) = v->z;
+      MatrixMultiply(voxelFromSurfaceRAS, v1, v2);
+      xv = V3_X(v2);
+      yv = V3_Y(v2);
+      zv = V3_Z(v2);
+#if 0
+      TransformSampleInverse
+      (transform, (float)xv, (float)yv, (float)zv, &xv2, &yv2, &zv2) ;
+#else
+      GCAMsampleMorph(gcam, xv, yv, zv, &xv2, &yv2, &zv2);
+#endif
+      MRIvoxelToSurfaceRAS(mri_dst, xv2, yv2, zv2, &xs, &ys, &zs);
+      v->x = xs;
+      v->y = ys;
+      v->z = zs;
+    }
+    mrisComputeSurfaceDimensions(mris);
+    // save the volume information from dst
+    getVolGeom(mri_dst, &mris->vg);
+    VectorFree(&v1);
+    VectorFree(&v2);
+    MatrixFree(&voxelFromSurfaceRAS);
+    if (dstNotGiven) {
+      MRIfree(&mri_dst);
+    }
+    if (srcNotGiven) {
+      MRIfree(&mri);
+    }
+  }
+  else {
+    lta = (LTA *)(transform->xform);
+    if (!mri_dst) {
+      dstNotGiven = 1;
+    }
+
+    // depend on the type of transform you have to handle differently
+    //
+    //       orig              ------>      RAS   c_(ras) != 0
+    //        |                              |
+    //        |                              | identity
+    //        V                              V
+    //    conformed vol        ------>      RAS   c_(ras) != 0
+    //        |                              |
+    //        | identity                     | surfaceRASFromRAS
+    //        V                              V
+    //    conformed vol        ------>   surfaceRAS  c_(ras) = 0
+    //
+    // given a volume transform you have to create a surfaceRAS transform
+    //
+    // Note that vertices are given by surfaceRAS coordinates
+    //
+    // RAS-to-RAS transform
+    //                  orig                 dst
+    //    surfaceRAS--->RAS --(ras-to-ras)-->RAS -->surfaceRAS
+    //
+    // VOX-to-Vox transform
+    //
+    //    surfaceRAS--->Vox---(vox-to-vox)-->Vox -->surfaceRAS
+    //
+    //
+    if (lta->num_xforms > 1) {
+      ErrorExit(ERROR_BADPARM, "we cannot handle multiple transforms\n");
+    }
+    if (lta->num_xforms == 0) {
+      ErrorExit(ERROR_BADPARM, "transform does not have transform ;-) \n");
+    }
+
+    // if volumes are not given, then try to get them from transform
+    lt = &lta->xforms[0];
+
+    /////////////////////////////////////////////////////////////////////
+    // The following rather fragile treatment about volume info is
+    // that lta may or may not store the valid src and dst info
+    // when the transform was created
+    // Another problem is due to the convention of conformed volume
+    ////////////////////////////////////////////////////////////////////
+    //////////////////////////////////////////////////////////////////
+    // check the c_ras values for mri side
+    if (lta->type == LINEAR_RAS_TO_RAS) {
+      if (mri && lt->src.valid == 1) {
+        if (!(FZERO(lt->src.c_r - lt->src.c_r) && FZERO(lt->src.c_a - lt->src.c_a) &&
+              FZERO(lt->src.c_s - lt->src.c_s))) {
+          fprintf(stderr, "WARNING:****************************************************\n");
+          fprintf(stderr,
+                  "WARNING: c_(ras) values are not equal for the "
+                  "input volume and the transform src.\n");
+          fprintf(stderr, "WARNING: The transformed surface position may be shifted.\n");
+          fprintf(stderr, "WARNING:***************************************************\n");
+        }
+      }
+      if (mri && mris->vg.valid == 1) {
+        if (!(FZERO(mris->vg.c_r - mris->vg.c_r) && FZERO(mris->vg.c_a - mris->vg.c_a) &&
+              FZERO(mris->vg.c_s - mris->vg.c_s))) {
+          fprintf(stderr,
+                  "WARNING:***************************"
+                  "******************************\n");
+          fprintf(stderr,
+                  "WARNING: c_(ras) values are not equal for the input volume "
+                  "and the surface stored volume.\n");
+          fprintf(stderr, "WARNING: The transformed surface position may be shifted.\n");
+          fprintf(stderr,
+                  "WARNING:*******************************"
+                  "**************************\n");
+        }
+      }
+    }
+
+    // if mri is not given, then use the one stored in the transform
+    if (!mri && lt->src.valid == 1) {
+      srcPresent = 0;
+      fprintf(stderr, "INFO:try to get src info from transform.\n");
+      mri = MRIallocHeader(lt->src.width, lt->src.height, lt->src.depth, MRI_UCHAR, 1);
+      mri->x_r = lt->src.x_r;
+      mri->y_r = lt->src.y_r;
+      mri->z_r = lt->src.z_r;
+      mri->c_r = lt->src.c_r;
+      mri->x_a = lt->src.x_a;
+      mri->y_a = lt->src.y_a;
+      mri->z_a = lt->src.z_a;
+      mri->c_a = lt->src.c_a;
+      mri->x_s = lt->src.x_s;
+      mri->y_s = lt->src.y_s;
+      mri->z_s = lt->src.z_s;
+      mri->c_s = lt->src.c_s;
+      mri->xsize = lt->src.xsize;
+      mri->ysize = lt->src.ysize;
+      mri->zsize = lt->src.zsize;
+      mri->ras_good_flag = 1;
+      MRIreInitCache(mri);
+    }
+    // if mri is not given and transform does not
+    // have it, get it from the surface
+    else if (!mri && mris->vg.valid == 1) {
+      fprintf(stderr, "INFO:try to get src info from the surface.\n");
+      mri = MRIallocHeader(mris->vg.width, mris->vg.height, mris->vg.depth, MRI_UCHAR, 1);
+      mri->x_r = mris->vg.x_r;
+      mri->y_r = mris->vg.y_r;
+      mri->z_r = mris->vg.z_r;
+      mri->c_r = mris->vg.c_r;
+      mri->x_a = mris->vg.x_a;
+      mri->y_a = mris->vg.y_a;
+      mri->z_a = mris->vg.z_a;
+      mri->c_a = mris->vg.c_a;
+      mri->x_s = mris->vg.x_s;
+      mri->y_s = mris->vg.y_s;
+      mri->z_s = mris->vg.z_s;
+      mri->c_s = mris->vg.c_s;
+      mri->xsize = mris->vg.xsize;
+      mri->ysize = mris->vg.ysize;
+      mri->zsize = mris->vg.zsize;
+      mri->ras_good_flag = 1;
+      MRIreInitCache(mri);
+    }
+    else if (!mri) {
+      error = 1;
+      strcpy(errMsg, "When mri == NULL, the transform must have the valid src info.\n");
+      goto mristransform_cleanup;
+    }
+    ///////////////////////////////////////////////////////////////////////////
+    // mri_dst side
+    // Note: if mri_dst is not given, override the one stored in the transform
+    if (!mri_dst && lt->dst.valid == 1) {
+      dstPresent = 0;
+      if (Gdiag & DIAG_SHOW && DIAG_VERBOSE_ON) {
+        fprintf(stderr, "INFO: Get dst info from transform.\n");
+      }
+      lt = &lta->xforms[0];
+      mri_dst = MRIallocHeader(lt->dst.width, lt->dst.height, lt->dst.depth, MRI_UCHAR, 1);
+      mri_dst->x_r = lt->dst.x_r;
+      mri_dst->y_r = lt->dst.y_r;
+      mri_dst->z_r = lt->dst.z_r;
+      mri_dst->c_r = lt->dst.c_r;
+      mri_dst->x_a = lt->dst.x_a;
+      mri_dst->y_a = lt->dst.y_a;
+      mri_dst->z_a = lt->dst.z_a;
+      mri_dst->c_a = lt->dst.c_a;
+      mri_dst->x_s = lt->dst.x_s;
+      mri_dst->y_s = lt->dst.y_s;
+      mri_dst->z_s = lt->dst.z_s;
+      mri_dst->c_s = lt->dst.c_s;
+      mri_dst->xsize = lt->dst.xsize;
+      mri_dst->ysize = lt->dst.ysize;
+      mri_dst->zsize = lt->dst.zsize;
+      mri_dst->ras_good_flag = 1;
+    }
+    else if (!mri_dst) {
+      fprintf(stderr, "WARNING:*********************************************************\n");
+      fprintf(stderr, "WARNING: transform does not have valid destination volume.       \n");
+      fprintf(stderr, "WARNING: The standard CORONAL volume with c_(ras) = 0 is assumed.\n");
+      fprintf(stderr, "WARNING:*********************************************************\n");
+      mri_dst = MRIallocHeader(lt->dst.width, lt->dst.height, lt->dst.depth, MRI_UCHAR, 1);
+      mri_dst->x_r = -1;
+      mri_dst->y_r = 0;
+      mri_dst->z_r = 0;
+      mri_dst->c_r = 0;
+      mri_dst->x_a = 0;
+      mri_dst->y_a = 0;
+      mri_dst->z_a = 1;
+      mri_dst->c_a = 0;
+      mri_dst->x_s = 0;
+      mri_dst->y_s = -1;
+      mri_dst->z_s = 0;
+      mri_dst->c_s = 0;
+      mri_dst->xsize = 1;
+      mri_dst->ysize = 1;
+      mri_dst->zsize = 1;
+      mri_dst->ras_good_flag = 1;
+    }
+    // WATCH /////////////////////////////////////////////////////////
+    // When dst is not given, our convention is that the
+    // dst is conformed and thus we need
+    // to modify mri_dst to be coronal conformed volume
+    // (even though transform target was not,
+    // as in the case of MNI average_305 being non-coronal, non-conformed
+    //////////////////////////////////////////////////////////////////
+    if (dstNotGiven) {
+      // assuming mri is conformed
+      mri_dst->width = mri->width;
+      mri_dst->height = mri->height;
+      mri_dst->depth = mri->depth;
+      mri_dst->xsize = mri->xsize;
+      mri_dst->ysize = mri->ysize;
+      mri_dst->zsize = mri->zsize;
+      mri_dst->x_r = -1;
+      mri_dst->y_r = 0;
+      mri_dst->z_r = 0;
+      mri_dst->x_a = 0;
+      mri_dst->y_a = 0;
+      mri_dst->z_a = 1;
+      mri_dst->x_s = 0;
+      mri_dst->y_s = -1;
+      mri_dst->z_s = 0;
+      // this means that we only retain c_ras info
+    }
+    // you must reinitialise cache
+    MRIreInitCache(mri_dst);
+
+    //  MatrixPrint(stderr, mri_dst->i_to_r__);
+
+    ////////////////////////////////////////////////////////////////
+    // Now we can calculate
+    if (lta->type == LINEAR_RAS_TO_RAS) {
+      //  we follow the right hand side of the map
+      //
+      //    conformed -----> surfaceRAS (c_ras = 0)
+      //       |              |
+      //       V              V
+      //    conformed -----> RAS (c_ras != 0)
+      //       |              |
+      //       |             xfm
+      //       V              v
+      //  conformed dst ---> RAS (c_ras != 0)
+      //       |              |
+      //       V              V
+      //  conformed dst ---> surfaceRAS (c_ras = 0)
+      //
+      if (mris->useRealRAS) {
+        surfaceRASFromSurfaceRAS = MatrixCopy(lta->xforms[0].m_L, NULL);
+      }
+      else {
+        RASFromSurfaceRAS = RASFromSurfaceRAS_(mri);      // needs only c_(ras) info
+        surfaceRASFromRAS = surfaceRASFromRAS_(mri_dst);  // need only c_(ras) info
+        m = MatrixMultiply(lta->xforms[0].m_L, RASFromSurfaceRAS, NULL);
+        surfaceRASFromSurfaceRAS = MatrixMultiply(surfaceRASFromRAS, m, NULL);
+      }
+    }
+    else if (lta->type == LINEAR_VOX_TO_VOX) {
+      if (mri->width != mri_dst->width || mri->height != mri_dst->height || mri->depth != mri_dst->depth) {
+        fprintf(stderr, "WARNING:********************************************************\n");
+        fprintf(stderr, "WARNING:voxel-to-voxel transform must have the same volume sizes.\n");
+        fprintf(stderr,
+                "WARNING:You gave src (%d, %dm, %d) vs. dst (%d, %d, %d).\n",
+                mri->width,
+                mri->height,
+                mri->depth,
+                mri_dst->width,
+                mri_dst->height,
+                mri_dst->depth);
+        fprintf(stderr, "WARNING:The result of this transform is most likely wrong.\n");
+        fprintf(stderr, "WARNING:********************************************************\n");
+      }
+      voxelFromSurfaceRAS = voxelFromSurfaceRAS_(mri);
+      surfaceRASFromVoxel = surfaceRASFromVoxel_(mri_dst);
+      //    MatrixPrint(stderr,  surfaceRASFromVoxel);
+      m = MatrixMultiply(lta->xforms[0].m_L, voxelFromSurfaceRAS, NULL);
+      surfaceRASFromSurfaceRAS = MatrixMultiply(surfaceRASFromVoxel, m, NULL);
+    }
+    // now apply the transform
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      // transform vertex point to actual voxel point
+      TransformWithMatrix(surfaceRASFromSurfaceRAS, v->x, v->y, v->z, &xw, &yw, &zw);
+      v->x = xw;
+      v->y = yw;
+      v->z = zw;
+    }
+    mrisComputeSurfaceDimensions(mris);
+    // save the volume information from dst
+    getVolGeom(mri_dst, &mris->vg);
+
+  mristransform_cleanup:
+    // free memory //////////////////////////////
+    if (RASFromSurfaceRAS) {
+      MatrixFree(&RASFromSurfaceRAS);
+    }
+    if (surfaceRASFromRAS) {
+      MatrixFree(&surfaceRASFromRAS);
+    }
+    if (m) {
+      MatrixFree(&m);
+    }
+    if (voxelFromSurfaceRAS) {
+      MatrixFree(&voxelFromSurfaceRAS);
+    }
+    if (surfaceRASFromVoxel) {
+      MatrixFree(&surfaceRASFromVoxel);
+    }
+    if (surfaceRASFromSurfaceRAS) {
+      MatrixFree(&surfaceRASFromSurfaceRAS);
+    }
+    if (!srcPresent && mri) {
+      MRIfree(&mri);
+    }
+    if (!dstPresent && mri_dst) {
+      MRIfree(&mri_dst);
+    }
+  }
+
+  if (error) {
+    ErrorExit(ERROR_BADPARM, errMsg);
+    return -1;  // just to satisfy compiler
+  }
+  else {
+    return (NO_ERROR);
+  }
+}
+
+
+float mrisComputeArea(MRIS *mris, int fac, int n)
+{
+  int n0, n1;
+  face_type *f;
+  float v0[3], v1[3], d1, d2, d3, dot, area;
+
+  n0 = (n == 0) ? VERTICES_PER_FACE - 1 : n - 1;
+  n1 = (n == VERTICES_PER_FACE - 1) ? 0 : n + 1;
+
+  f = &mris->faces[fac];
+
+  v0[0] = mris->vertices[f->v[n]].x - mris->vertices[f->v[n0]].x;
+  v0[1] = mris->vertices[f->v[n]].y - mris->vertices[f->v[n0]].y;
+  v0[2] = mris->vertices[f->v[n]].z - mris->vertices[f->v[n0]].z;
+
+  v1[0] = mris->vertices[f->v[n1]].x - mris->vertices[f->v[n]].x;
+  v1[1] = mris->vertices[f->v[n1]].y - mris->vertices[f->v[n]].y;
+  v1[2] = mris->vertices[f->v[n1]].z - mris->vertices[f->v[n]].z;
+
+  d1 = -v1[1] * v0[2] + v0[1] * v1[2];
+  d2 =  v1[0] * v0[2] - v0[0] * v1[2];
+  d3 = -v1[0] * v0[1] + v0[0] * v1[1];
+
+  dot = mris->vertices[f->v[n]].x * d1 + mris->vertices[f->v[n]].y * d2 + mris->vertices[f->v[n]].z * d3;
+
+  if (dot < 0.0f) /* not in same direction, area < 0 and reverse n */
+  {
+    area = -sqrt(d1 * d1 + d2 * d2 + d3 * d3);
+  }
+  else {
+    area = sqrt(d1 * d1 + d2 * d2 + d3 * d3);
+  }
+
+  return area;
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISzeroNegativeAreas(MRIS *mris)
+{
+  int vno, fno;
+  VERTEX *v;
+  FACE *face;
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    if (v->area < 0) {
+      v->area = 0;
+    }
+  }
+  for (fno = 0; fno < mris->nfaces; fno++) {
+    face = &mris->faces[fno];
+    if (face->ripflag) {
+      continue;
+    }
+    if (face->area < 0.0f) {
+      face->area = 0.0f;
+    }
+  }
+  return (NO_ERROR);
+}
+
+/*! ----------------------------------------------------
+  \fn int MRISclearCurvature(MRIS *mris)
+  \brief sets v->curv=0 for unripped vertices.
+  ------------------------------------------------------*/
+int MRISclearCurvature(MRIS *mris)
+{
+  int vno;
+  VERTEX *v;
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->curv = 0;
+  }
+  return (NO_ERROR);
+}
+
+int MRISrectifyCurvature(MRIS *mris)
+{
+  int vno;
+  VERTEX *v;
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    v->curv = fabs(v->curv);
+  }
+  mrisComputeCurvatureMinMax(mris);
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISnonmaxSuppress(MRIS *mris)
+{
+  double du, dv, up1, um1, vp1, vm1, src, dx, dy, dz, fp1, fm1, mag;
+  VERTEX *v;
+  int vno;
+  float x, y, z, e1x, e1y, e1z, e2x, e2y, e2z, ux, uy, uz, vx, vy, vz;
+  MRI_SP *mrisp, *mrisp_blur;
+  double d_dist = D_DIST * mris->avg_vertex_dist;
+
+  mrisp = MRIStoParameterization(mris, NULL, 1, 0);
+  mrisp_blur = MRISPblur(mrisp, NULL, 20.0, 0);
+  mrisComputeTangentPlanes(mris);
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+
+    e1x = v->e1x;
+    e1y = v->e1y;
+    e1z = v->e1z;
+    e2x = v->e2x;
+    e2y = v->e2y;
+    e2z = v->e2z;
+    x = v->x;
+    y = v->y;
+    z = v->z;
+    src = MRISPfunctionVal(mrisp, mris, x, y, z, 0);
+
+    /* now compute gradient of template w.r.t. a change in vertex position */
+
+    /*
+      sample the curvature functions along the tangent plane axes and
+      compute the derivates using them.
+    */
+    ux = 10 * e1x;
+    uy = 10 * e1y;
+    uz = 10 * e1z;
+    vx = 10 * e2x;
+    vy = 10 * e2y;
+    vz = 10 * e2z;
+    ux = e1x * d_dist;
+    uy = e1y * d_dist;
+    uz = e1z * d_dist;
+    vx = e2x * d_dist;
+    vy = e2y * d_dist;
+    vz = e2z * d_dist;
+
+    /* compute gradient usnig blurred image */
+    up1 = MRISPfunctionVal(mrisp_blur, mris, x + ux, y + uy, z + uz, 0);
+    um1 = MRISPfunctionVal(mrisp_blur, mris, x - ux, y - uy, z - uz, 0);
+    vp1 = MRISPfunctionVal(mrisp_blur, mris, x + vx, y + vy, z + vz, 0);
+    vm1 = MRISPfunctionVal(mrisp_blur, mris, x - vx, y - vy, z - vz, 0);
+    du = (up1 - um1) / (2 * d_dist);
+    dv = (vp1 - vm1) / (2 * d_dist);
+
+    /* calculate curvature gradient */
+    dx = (du * e1x + dv * e2x);
+    dy = (du * e1y + dv * e2y);
+    dz = (du * e1z + dv * e2z);
+    mag = sqrt(dx * dx + dy * dy + dz * dz);
+    if (FZERO(mag)) /* zero gradient */
+    {
+      v->curv = 0;
+    }
+    else {
+      mag *= .1;
+      dx = dx / mag;
+      dy = dy / mag;
+      dz = dz / mag;
+    }
+    fp1 = MRISPfunctionVal(mrisp, mris, x + dx, y + dy, z + dz, 0);
+    fm1 = MRISPfunctionVal(mrisp, mris, x - dx, y - dy, z - dz, 0);
+
+    if ((src >= fp1) && (src >= fm1)) /* local max */
+    {
+      v->curv = 1;
+    }
+    else if ((src <= fp1) && (src <= fm1)) /* local min */
+    {
+      v->curv = -1;
+    }
+    else {
+      v->curv = 0;
+    }
+    v->curv = src;
+  }
+  MRISPfree(&mrisp);
+  return (NO_ERROR);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISscaleCurvatures(MRIS *mris, float min_curv, float max_curv)
+{
+  double old_min_curv, old_max_curv, mean, scale;
+  int vno, vtotal;
+  VERTEX *v;
+
+  old_min_curv = 100000.0;
+  old_max_curv = -100000.0;
+  for (mean = 0.0, vtotal = vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    vtotal++;
+    mean += v->curv;
+    if (v->curv > old_max_curv) {
+      old_max_curv = v->curv;
+    }
+    if (v->curv < old_min_curv) {
+      old_min_curv = v->curv;
+    }
+  }
+
+  mean /= (double)vtotal;
+  scale = (max_curv - min_curv) / (old_max_curv - old_min_curv);
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->curv = (v->curv - mean) * scale + mean;
+  }
+
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  File Format is:
+
+  name x y z
+  ------------------------------------------------------*/
+#define K_A 0.4f
+static float kernel[] = {K_A, 0.25f, 0.25f - K_A / 2.0f};
+int MRISsmoothCurvatures(MRIS *mris, int niterations)
+{
+  int vno, i, vn;
+  double g, H, norm;
+
+  for (i = 0; i < niterations; i++) {
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      VERTEX_TOPOLOGY const * const vertext = &mris->vertices_topology[vno];
+      VERTEX                * const vertex  = &mris->vertices         [vno];
+      if (vertex->ripflag) {
+        continue;
+      }
+      H = kernel[0] * vertex->H;
+      g = kernel[1];
+      for (vn = 0; vn < vertext->vnum; vn++) {
+        H += g * mris->vertices[vertext->v[vn]].H;
+      }
+      g = kernel[2];
+      for (; vn < vertext->v2num; vn++) {
+        H += g * mris->vertices[vertext->v[vn]].H;
+      }
+      norm = kernel[0] + vertext->vnum * kernel[1] + (vertext->v2num - vertext->vnum) * kernel[2];
+      vertex->d = H / norm;
+    }
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      VERTEX * const vertex = &mris->vertices[vno];
+      if (vertex->ripflag) {
+        continue;
+      }
+      vertex->H = vertex->d;
+    }
+  }
+
+  return (NO_ERROR);
+}
+
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISzeroMeanCurvature(MRIS *mris)
+{
+  double mean;
+  int vno, vtotal;
+  VERTEX *v;
+
+  for (mean = 0.0f, vtotal = vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    mean += v->curv;
+    vtotal++;
+  }
+
+  mean /= (double)vtotal;
+
+  if (Gdiag & DIAG_SHOW && DIAG_VERBOSE_ON) {
+    fprintf(stdout, "curvature mean = %2.3f\n", mean);
+  }
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->curv -= mean;
+  }
+
+  mrisComputeCurvatureMinMax(mris);
+  return (NO_ERROR);
+}
+/*-----------------------------------------------------
+int MRISnormalizeCurvature(MRIS *mris, int which_norm)
+Normalize the curvatures so they have unit standard deviation.
+which_norm can be NORM_MEDIAN or NORM_MEAN. NORM_MEAN results
+in the usual stddev computation. NORM_MEDIAN uses the median
+instead of the mean for the stddev computation.
+  ------------------------------------------------------*/
+int MRISnormalizeCurvature(MRIS *mris, int which_norm)
+{
+  double mean, var, std, median;
+  int vno, vtotal;
+  VERTEX *v;
+
+  if (which_norm == NORM_NONE) {
+    return (NO_ERROR);
+  }
+
+  if (which_norm == NORM_MEDIAN) {
+    float *curvs;
+    curvs = (float *)calloc(mris->nvertices, sizeof(float));
+    if (curvs == NULL) {
+      ErrorExit(ERROR_NOMEMORY, "MRISnormalize(MEDIAN): couldn't alloc array");
+    }
+
+    MRISextractCurvatureVector(mris, curvs);
+    qsort(curvs, mris->nvertices, sizeof(curvs[0]), mris_sort_compare_float);
+    median = curvs[mris->nvertices / 2];
+    free(curvs);
+    for (var = 0.0f, vtotal = vno = 0; vno < mris->nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      std = (v->curv - median);
+      vtotal++;
+      var += std * std;
+    }
+
+    var /= (double)vtotal;
+    std = sqrt(var);
+
+    //    if (Gdiag & DIAG_SHOW && DIAG_VERBOSE_ON)
+    fprintf(stdout, "curvature median = %2.3f, std = %2.3f\n", median, std);
+
+    // now normalize the curvatures so they have unit standard deviation
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      v->curv = (v->curv - median) / std;
+    }
+  }
+  else  // normalize mean
+  {
+    for (mean = 0.0f, vtotal = vno = 0; vno < mris->nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      mean += v->curv;
+      vtotal++;
+    }
+
+    mean /= (double)vtotal;
+
+    for (var = 0.0f, vno = 0; vno < mris->nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      std = v->curv - mean;
+      var += std * std;
+    }
+
+    var /= (double)vtotal;
+    std = sqrt(var);
+
+    //    if (Gdiag & DIAG_SHOW && DIAG_VERBOSE_ON)
+    fprintf(stdout, "curvature mean = %2.3f, std = %2.3f\n", mean, std);
+
+    /* now normalize the curvatures so they have unit standard deviation, but
+       leave the mean alone */
+    for (vno = 0; vno < mris->nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      v->curv = (v->curv - mean) / std /* + mean*/;
+    }
+  }
+  mrisComputeCurvatureMinMax(mris);
+  return (NO_ERROR);
+}
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISnormalizeCurvatureVariance(MRIS *mris)
+{
+  double mean, var, std;
+  int vno, vtotal;
+  VERTEX *v;
+
+  for (mean = 0.0f, vtotal = vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    mean += v->curv;
+    vtotal++;
+  }
+
+  mean /= (double)vtotal;
+
+  for (var = 0.0f, vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    std = v->curv - mean;
+    var += std * std;
+  }
+
+  var /= (double)vtotal;
+  std = sqrt(var);
+
+  if (Gdiag & DIAG_SHOW && DIAG_VERBOSE_ON) {
+    fprintf(stdout, "curvature mean = %2.3f, std = %2.3f\n", mean, std);
+  }
+
+  /* now normalize the curvatures so they have unit standard deviation, but
+     leave the mean alone */
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->curv = (v->curv - mean) / std + mean;
+  }
+
+  mrisComputeCurvatureMinMax(mris);
+  return (NO_ERROR);
+}
+
+
+
+int mrisCountCompressed(MRIS *mris, double min_dist)
+{
+  int vno, n, num;
+  double d, dx, dy, dz;
+
+  for (vno = num = 0; vno < mris->nvertices; vno++) {
+    VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+    VERTEX          const * const v  = &mris->vertices         [vno];
+    if (vno == Gdiag_no) {
+      DiagBreak();
+    }
+    if (v->ripflag) {
+      continue;
+    }
+    for (n = 0; n < vt->vnum; n++) {
+      VERTEX const * const vn = &mris->vertices[vt->v[n]];
+      if (vn->ripflag) {
+        continue;
+      }
+      dx = vn->x - v->x;
+      dy = vn->y - v->y;
+      dz = vn->z - v->z;
+      d = sqrt(dx * dx + dy * dy + dz * dz);
+      if (d < min_dist) {
+        num++;
+      }
+    }
+  }
+  return (num);
+}
+
+
+//=========================================================================
+// Change the positions of the vertices
+//=========================================================================
+
+int MRISrestoreOldPositions(MRIS *mris)
+{
+  return (MRISrestoreVertexPositions(mris, TMP_VERTICES));
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISstoreCurrentPositions(MRIS *mris)
+{
+  return (MRISsaveVertexPositions(mris, TMP_VERTICES));
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISstoreMeanCurvature(MRIS *mris)
+{
+  int vno, nvertices;
+  VERTEX *v;
+
+  nvertices = mris->nvertices;
+  for (vno = 0; vno < nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->H = v->curv;
+  }
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISstoreMetricProperties(MRIS *mris)
+{
+  int vno, nvertices, fno, ano, n;
+  FACE *f;
+
+#if 0
+  MRIScomputeNormals(mris);              /* update vertex areas */
+  MRIScomputeTriangleProperties(mris) ;  /* update triangle properties */
+#endif
+  nvertices = mris->nvertices;
+  for (vno = 0; vno < nvertices; vno++) {
+    VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+    VERTEX                * const v  = &mris->vertices         [vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->origarea = v->area;
+#if 1
+    if (v->dist && v->dist_orig)
+      for (n = 0; n < vt->vtotal; n++) {
+        v->dist_orig[n] = v->dist[n];
+      }
+#endif
+  }
+  for (fno = 0; fno < mris->nfaces; fno++) {
+    f = &mris->faces[fno];
+    if (f->ripflag) {
+      continue;
+    }
+    setFaceOrigArea(mris, fno, f->area);
+    for (ano = 0; ano < ANGLES_PER_TRIANGLE; ano++) {
+      f->orig_angle[ano] = f->angle[ano];
+    }
+  }
+  mris->orig_area = mris->total_area;
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISrestoreMetricProperties(MRIS *mris)
+{
+  int vno, nvertices, fno, ano, n;
+  FACE *f;
+
+  nvertices = mris->nvertices;
+  for (vno = 0; vno < nvertices; vno++) {
+    VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+    VERTEX                * const v  = &mris->vertices         [vno];
+    if (v->ripflag) {
+      continue;
+    }
+    v->area = v->origarea;
+    for (n = 0; n < vt->vtotal; n++) {
+      v->dist[n] = v->dist_orig[n];
+    }
+  }
+  for (fno = 0; fno < mris->nfaces; fno++) {
+    f = &mris->faces[fno];
+    if (f->ripflag) {
+      continue;
+    }
+    f->area = getFaceOrigArea(mris, fno);
+    for (ano = 0; ano < ANGLES_PER_TRIANGLE; ano++) {
+      f->angle[ano] = f->orig_angle[ano];
+    }
+  }
+  mris->total_area = mris->orig_area;
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISsaveVertexPositions(MRIS *mris, int which)
+{
+  int vno, nvertices;
+  VERTEX *v;
+
+  nvertices = mris->nvertices;
+  for (vno = 0; vno < nvertices; vno++) {
+    v = &mris->vertices[vno];
+#if 0
+    if (v->ripflag)
+    {
+      continue ;
+    }
+#endif
+    switch (which) {
+      case LAYERIV_VERTICES:
+        v->l4x = v->x;
+        v->l4y = v->y;
+        v->l4z = v->z;
+        break;
+      case TARGET_VERTICES:
+        v->targx = v->x;
+        v->targy = v->y;
+        v->targz = v->z;
+        break;
+      case WHITE_VERTICES:
+        v->whitex = v->x;
+        v->whitey = v->y;
+        v->whitez = v->z;
+        break;
+      case PIAL_VERTICES:
+        v->pialx = v->x;
+        v->pialy = v->y;
+        v->pialz = v->z;
+        break;
+      case INFLATED_VERTICES:
+        v->infx = v->x;
+        v->infy = v->y;
+        v->infz = v->z;
+        break;
+      case FLATTENED_VERTICES:
+        v->fx = v->x;
+        v->fy = v->y;
+        v->fz = v->z;
+        break;
+      case CANONICAL_VERTICES:
+        v->cx = v->x;
+        v->cy = v->y;
+        v->cz = v->z;
+        break;
+      case ORIGINAL_VERTICES:
+        v->origx = v->x; CHANGES_ORIG
+        v->origy = v->y;
+        v->origz = v->z;
+        break;
+      case TMP2_VERTICES:
+        v->tx2 = v->x;
+        v->ty2 = v->y;
+        v->tz2 = v->z;
+        break;
+      default:
+      case TMP_VERTICES:
+        v->tx = v->x;
+        v->ty = v->y;
+        v->tz = v->z;
+        break;
+    }
+  }
+  if (which == CANONICAL_VERTICES) {
+    MRIScomputeCanonicalCoordinates(mris);
+  }
+  return (NO_ERROR);
+}
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISsaveNormals(MRIS *mris, int which)
+{
+  int vno, nvertices;
+  VERTEX *v;
+
+  nvertices = mris->nvertices;
+  for (vno = 0; vno < nvertices; vno++) {
+    v = &mris->vertices[vno];
+    switch (which) {
+      case TMP_VERTICES:
+        v->tx = v->nx;
+        v->ty = v->ny;
+        v->tz = v->nz;
+        break;
+      case TMP2_VERTICES:
+        v->tx2 = v->nx;
+        v->ty2 = v->ny;
+        v->tz2 = v->nz;
+        break;
+      case PIAL_VERTICES:
+        v->pnx = v->nx;
+        v->pny = v->ny;
+        v->pnz = v->nz;
+        break;
+      case WHITE_VERTICES:
+        v->wnx = v->nx;
+        v->wny = v->ny;
+        v->wnz = v->nz;
+        break;
+      case INFLATED_VERTICES:
+      case FLATTENED_VERTICES:
+      case CANONICAL_VERTICES:
+      case ORIGINAL_VERTICES:
+      default:
+        ErrorExit(ERROR_BADPARM, "MRISsaveNormals: unsupported which %d", which);
+        break;
+    }
+  }
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISrestoreNormals(MRIS *mris, int which)
+{
+  int vno, nvertices;
+  VERTEX *v;
+
+  nvertices = mris->nvertices;
+  for (vno = 0; vno < nvertices; vno++) {
+    v = &mris->vertices[vno];
+    switch (which) {
+      case TMP_VERTICES:
+        v->nx = v->tx;
+        v->ny = v->ty;
+        v->nz = v->tz;
+        break;
+      case TMP2_VERTICES:
+        v->nx = v->tx2;
+        v->ny = v->ty2;
+        v->nz = v->tz2;
+        break;
+      case PIAL_VERTICES:
+        v->nx = v->pnx;
+        v->ny = v->pny;
+        v->nz = v->pnz;
+        break;
+      case WHITE_VERTICES:
+      case INFLATED_VERTICES:
+      case FLATTENED_VERTICES:
+      case CANONICAL_VERTICES:
+      case ORIGINAL_VERTICES:
+      default:
+        ErrorExit(ERROR_BADPARM, "MRISsaveNormals: unsupported which %d", which);
+        break;
+    }
+  }
+  return (NO_ERROR);
+}
+
+/*-----------------------------------------------------
+  Parameters:
+
+  Returns value:
+
+  Description
+  ------------------------------------------------------*/
+int MRISrestoreVertexPositions(MRIS *mris, int which)
+{
+  int vno, nvertices;
+  VERTEX *v;
+
+  nvertices = mris->nvertices;
+  for (vno = 0; vno < nvertices; vno++) {
+    v = &mris->vertices[vno];
+#if 0
+    if (v->ripflag)
+    {
+      continue ;
+    }
+#endif
+    switch (which) {
+      case TARGET_VERTICES:
+        v->x = v->targx;
+        v->y = v->targy;
+        v->z = v->targz;
+        break;
+      case WHITE_VERTICES:
+        v->x = v->whitex;
+        v->y = v->whitey;
+        v->z = v->whitez;
+        break;
+      case LAYERIV_VERTICES:
+        v->x = v->l4x;
+        v->y = v->l4y;
+        v->z = v->l4z;
+        break;
+      case PIAL_VERTICES:
+        v->x = v->pialx;
+        v->y = v->pialy;
+        v->z = v->pialz;
+        break;
+      case INFLATED_VERTICES:
+        v->x = v->infx;
+        v->y = v->infy;
+        v->z = v->infz;
+        break;
+      case FLATTENED_VERTICES:
+        v->x = v->fx;
+        v->y = v->fy;
+        v->z = v->fz;
+        break;
+      case CANONICAL_VERTICES:
+        v->x = v->cx;
+        v->y = v->cy;
+        v->z = v->cz;
+        break;
+      case ORIGINAL_VERTICES:
+        v->x = v->origx;
+        v->y = v->origy;
+        v->z = v->origz;
+        break;
+      case TMP2_VERTICES:
+        v->x = v->tx2;
+        v->y = v->ty2;
+        v->z = v->tz2;
+        break;
+      default:
+      case TMP_VERTICES:
+        v->x = v->tx;
+        v->y = v->ty;
+        v->z = v->tz;
+        break;
+    }
+  }
+  if (mris->status == MRIS_SPHERICAL_PATCH) {
+    for (vno = 0; vno < nvertices; vno++) {
+      v = &mris->vertices[vno];
+      if (v->ripflag) {
+        continue;
+      }
+      switch (which) {
+        case WHITE_VERTICES:
+          v->cx = v->whitex;
+          v->cy = v->whitey;
+          v->cz = v->whitez;
+          break;
+        case PIAL_VERTICES:
+          v->cx = v->pialx;
+          v->cy = v->pialy;
+          v->cz = v->pialz;
+          break;
+        case INFLATED_VERTICES:
+          v->cx = v->infx;
+          v->cy = v->infy;
+          v->cz = v->infz;
+          break;
+        case FLATTENED_VERTICES:
+          v->cx = v->fx;
+          v->cy = v->fy;
+          v->cz = v->fz;
+          break;
+        case CANONICAL_VERTICES:
+          v->cx = v->cx;
+          v->cy = v->cy;
+          v->cz = v->cz;
+          break;
+        case ORIGINAL_VERTICES:
+          v->cx = v->origx;
+          v->cy = v->origy;
+          v->cz = v->origz;
+          break;
+        case TMP2_VERTICES:
+          v->cx = v->tx2;
+          v->cy = v->ty2;
+          v->cz = v->tz2;
+          break;
+        default:
+        case TMP_VERTICES:
+          v->cx = v->tx;
+          v->cy = v->ty;
+          v->cz = v->tz;
+          break;
+      }
+    }
+  }
+  mrisComputeSurfaceDimensions(mris);
+  return (NO_ERROR);
 }
 
 
@@ -538,7 +2494,7 @@ float FACES_commonEdgeLength_find(MRIS *apmris, FACE *apFACE_I, FACE *apFACE_J)
 
   Description
   ------------------------------------------------------*/
-int mrisComputeTangentPlanes(MRI_SURFACE *mris)
+int mrisComputeTangentPlanes(MRIS *mris)
 {
   VECTOR *v_n, *v_e1, *v_e2, *v;
   int vno;
@@ -588,7 +2544,7 @@ int mrisComputeTangentPlanes(MRI_SURFACE *mris)
   return (NO_ERROR);
 }
 
-int MRISstoreTangentPlanes(MRI_SURFACE *mris, int which_vertices)
+int MRISstoreTangentPlanes(MRIS *mris, int which_vertices)
 {
   VERTEX *v;
   int vno;
@@ -609,7 +2565,7 @@ int MRISstoreTangentPlanes(MRI_SURFACE *mris, int which_vertices)
 }
 
 
-int mrisComputeOptimalPlane(MRI_SURFACE *mris,
+int mrisComputeOptimalPlane(MRIS *mris,
                             int *vertices,
                             int nvertices,
                             double *pxn,
@@ -698,7 +2654,7 @@ int mrisComputeOptimalPlane(MRI_SURFACE *mris,
   Description
   Compute the folding of the surface.
   ------------------------------------------------------*/
-double MRIScomputeFolding(MRI_SURFACE *mris)
+double MRIScomputeFolding(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -726,7 +2682,7 @@ double MRIScomputeFolding(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-double MRISpercentDistanceError(MRI_SURFACE *mris)
+double MRISpercentDistanceError(MRIS *mris)
 {
   int vno, n, nvertices;
   double dist_scale, pct, dist, odist, mean, mean_error;
@@ -795,7 +2751,7 @@ double MRISpercentDistanceError(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRIScomputeCanonicalCoordinates(MRI_SURFACE *mris)
+int MRIScomputeCanonicalCoordinates(MRIS *mris)
 {
   float theta, phi, r, d, x, y, z;
   VERTEX *v;
@@ -959,7 +2915,7 @@ int MRISvertexCoord2XYZ_double(VERTEX *v, int which, double *x, double *y, doubl
   Description
   Scale a surface anisotropically.
   ------------------------------------------------------*/
-int mrisCalculateCanonicalFaceCentroid(MRI_SURFACE *mris, int fno, float *px, float *py, float *pz)
+int mrisCalculateCanonicalFaceCentroid(MRIS *mris, int fno, float *px, float *py, float *pz)
 {
   float x, y, z;
   VERTEX *v0, *v1, *v2;
@@ -990,7 +2946,7 @@ int mrisCalculateCanonicalFaceCentroid(MRI_SURFACE *mris, int fno, float *px, fl
   Description
   Scale a surface anisotropically.
   ------------------------------------------------------*/
-static int mrisCalculateOriginalFaceCentroid(MRI_SURFACE *mris, int fno, float *px, float *py, float *pz)
+static int mrisCalculateOriginalFaceCentroid(MRIS *mris, int fno, float *px, float *py, float *pz)
 {
   float x, y, z;
   VERTEX *v0, *v1, *v2;
@@ -1021,7 +2977,7 @@ static int mrisCalculateOriginalFaceCentroid(MRI_SURFACE *mris, int fno, float *
   Description
   Scale a surface anisotropically.
   ------------------------------------------------------*/
-int mrisCalculateFaceCentroid(MRI_SURFACE *mris, int fno, float *px, float *py, float *pz)
+int mrisCalculateFaceCentroid(MRIS *mris, int fno, float *px, float *py, float *pz)
 {
   float x, y, z;
   VERTEX *v0, *v1, *v2;
@@ -1055,7 +3011,7 @@ int mrisCalculateFaceCentroid(MRI_SURFACE *mris, int fno, float *px, float *py, 
   of v.
   ------------------------------------------------------*/
 
-int mrisChooseFace(MRI_SURFACE *mris, MHT *mht, VERTEX *v)
+int mrisChooseFace(MRIS *mris, MHT *mht, VERTEX *v)
 {
   int fno, nfound, flist[1000], min_fno, i, j, total_found, *fptr;
   double dist, d;
@@ -1131,7 +3087,7 @@ int mrisChooseFace(MRI_SURFACE *mris, MHT *mht, VERTEX *v)
   Find the face in which v lies. If it lies in more than one
   face, return not found.
   ------------------------------------------------------*/
-int mrisFindUnambiguousFace(MRI_SURFACE *mris, MHT *mht, VERTEX *v, int *pnfound)
+int mrisFindUnambiguousFace(MRIS *mris, MHT *mht, VERTEX *v, int *pnfound)
 {
   int nfound, flist[1000], *fptr, total_found, i, j;
   double dist, d;
@@ -1204,7 +3160,7 @@ double MRISavgVetexRadius(MRIS *Surf, double *StdDev)
   Description
   ------------------------------------------------------*/
 #include "stats.h"
-int MRISsampleStatVolume(MRI_SURFACE *mris, void *vsv, int time_point, int coords)
+int MRISsampleStatVolume(MRIS *mris, void *vsv, int time_point, int coords)
 {
   VERTEX *v;
   int vno, xv, yv, zv, width, height, depth;
@@ -1262,7 +3218,7 @@ int MRISsampleStatVolume(MRI_SURFACE *mris, void *vsv, int time_point, int coord
 
   Description
   ------------------------------------------------------*/
-double MRIScomputeFaceAreaStats(MRI_SURFACE *mris, double *psigma, double *pmin, double *pmax)
+double MRIScomputeFaceAreaStats(MRIS *mris, double *psigma, double *pmin, double *pmax)
 {
   double total_area, mean, var, nf, sigma, min_area, max_area, area, area_scale;
   int fno;
@@ -1313,7 +3269,7 @@ double MRIScomputeFaceAreaStats(MRI_SURFACE *mris, double *psigma, double *pmin,
   Description
   ------------------------------------------------------*/
 double MRIScomputeVertexSpacingStats(
-    MRI_SURFACE *mris, double *psigma, double *pmin, double *pmax, int *pvno, int *pvno2, int which_vertices)
+    MRIS *mris, double *psigma, double *pmin, double *pmax, int *pvno, int *pvno2, int which_vertices)
 {
   double total_dist, mean, var, nv, dist, sigma, min_dist, max_dist, dist_scale;
   int vno, n;
@@ -1404,7 +3360,7 @@ double MRIScomputeVertexSpacingStats(
   Description
   ------------------------------------------------------*/
 double MRIScomputeTotalVertexSpacingStats(
-    MRI_SURFACE *mris, double *psigma, double *pmin, double *pmax, int *pvno, int *pvno2)
+    MRIS *mris, double *psigma, double *pmin, double *pmax, int *pvno, int *pvno2)
 {
   double total_dist, mean, var, nv, dist, sigma, min_dist, max_dist, dist_scale;
   int vno, n;
@@ -1473,15 +3429,15 @@ double MRIScomputeTotalVertexSpacingStats(
   Expand the list of neighbors of each vertex, reallocating
   the v->v array to hold the expanded list.
   ------------------------------------------------------*/
-static void MRISsetNeighborhoodSizeAndDistWkr(MRI_SURFACE *mris, int nsize);
+static void MRISsetNeighborhoodSizeAndDistWkr(MRIS *mris, int nsize);
 
-void MRISsetNeighborhoodSizeAndDist(MRI_SURFACE *mris, int nsize) {
+void MRISsetNeighborhoodSizeAndDist(MRIS *mris, int nsize) {
   mrisCheckVertexFaceTopology(mris);
   MRISsetNeighborhoodSizeAndDistWkr(mris, nsize);
   mrisCheckVertexFaceTopology(mris);
 }
 
-static void MRISsetNeighborhoodSizeAndDistWkr(MRI_SURFACE *mris, int nsize)
+static void MRISsetNeighborhoodSizeAndDistWkr(MRIS *mris, int nsize)
 {
   cheapAssert(1 <= nsize && nsize < 4);
   int vno, niter, ntotal, vtotal;
@@ -1728,7 +3684,7 @@ static void MRISsetNeighborhoodSizeAndDistWkr(MRI_SURFACE *mris, int nsize)
 
 
 int MRISsampleFaceCoordsCanonical(
-    MHT *mht, MRI_SURFACE *mris, float x, float y, float z, int which, float *px, float *py, float *pz)
+    MHT *mht, MRIS *mris, float x, float y, float z, int which, float *px, float *py, float *pz)
 {
   float xv, yv, zv;
   double lambda[3], fdist, norm;
@@ -1766,7 +3722,7 @@ int MRISsampleFaceCoordsCanonical(
   return (ret);
 }
 
-int IsMRISselfIntersecting(MRI_SURFACE *mris)
+int IsMRISselfIntersecting(MRIS *mris)
 {
   MRIS_HASH_TABLE *mht;
   int fno;
@@ -1785,7 +3741,7 @@ int IsMRISselfIntersecting(MRI_SURFACE *mris)
 }
 
 
-int face_barycentric_coords(MRI_SURFACE const *mris,
+int face_barycentric_coords(MRIS const *mris,
                             int fno,
                             int which_vertices,
                             double cx,
@@ -1904,7 +3860,7 @@ int face_barycentric_coords(MRI_SURFACE const *mris,
   return (ret);
 }
 
-int MRISsampleFaceCoords(MRI_SURFACE *mris,
+int MRISsampleFaceCoords(MRIS *mris,
                          int fno,
                          double x,
                          double y,
@@ -1942,7 +3898,7 @@ int MRISsampleFaceCoords(MRI_SURFACE *mris,
 }
 
 double MRISsampleFace(
-    MRI_SURFACE *mris, int fno, int which, double x, double y, double z, double val0, double val1, double val2)
+    MRIS *mris, int fno, int which, double x, double y, double z, double val0, double val1, double val2)
 {
   float xv, yv, zv;
   double lambda[3], val;
@@ -1975,7 +3931,7 @@ double MRISsampleFace(
 
   return (val);
 }
-int MRISsampleFaceNormal(MRI_SURFACE *mris, int fno, double x, double y, double z, float *px, float *py, float *pz)
+int MRISsampleFaceNormal(MRIS *mris, int fno, double x, double y, double z, float *px, float *py, float *pz)
 {
   float d, dtotal, dx, dy, dz, xc, yc, zc;
   int n;
@@ -2027,7 +3983,7 @@ int MRISsampleFaceNormal(MRI_SURFACE *mris, int fno, double x, double y, double 
 
   Description
   ------------------------------------------------------*/
-int MRISfindClosestVertex(MRI_SURFACE *mris, float x, float y, float z, float *dmin, int which_vertices)
+int MRISfindClosestVertex(MRIS *mris, float x, float y, float z, float *dmin, int which_vertices)
 {
   int vno, min_v = -1;
   VERTEX *v;
@@ -2077,7 +4033,7 @@ int MRISfindClosestVertex(MRI_SURFACE *mris, float x, float y, float z, float *d
 
   Description
   ------------------------------------------------------*/
-int MRISfindClosestOriginalVertex(MRI_SURFACE *mris, float x, float y, float z)
+int MRISfindClosestOriginalVertex(MRIS *mris, float x, float y, float z)
 {
   int vno, min_v = -1;
   VERTEX *v;
@@ -2111,7 +4067,7 @@ int MRISfindClosestOriginalVertex(MRI_SURFACE *mris, float x, float y, float z)
 
   Description
   ------------------------------------------------------*/
-int MRISfindClosestCanonicalVertex(MRI_SURFACE *mris, float x, float y, float z)
+int MRISfindClosestCanonicalVertex(MRIS *mris, float x, float y, float z)
 {
   int vno, min_v = -1;
   VERTEX *v;
@@ -2143,7 +4099,7 @@ int MRISfindClosestCanonicalVertex(MRI_SURFACE *mris, float x, float y, float z)
 
   Description
   ------------------------------------------------------*/
-int MRISfindClosestWhiteVertex(MRI_SURFACE *mris, float x, float y, float z)
+int MRISfindClosestWhiteVertex(MRIS *mris, float x, float y, float z)
 {
   int vno, min_v = -1;
   VERTEX *v;
@@ -2175,7 +4131,7 @@ int MRISfindClosestWhiteVertex(MRI_SURFACE *mris, float x, float y, float z)
 
   Description
   ------------------------------------------------------*/
-int mrisFindAllOverlappingFaces(MRI_SURFACE *mris, MHT *mht, int fno, int *flist)
+int mrisFindAllOverlappingFaces(MRIS *mris, MHT *mht, int fno, int *flist)
 {
   double x0, x1, y0, y1, z0, z1, x, y, z;
   int i, n, m, total_found, all_faces[1000000], nfaces;
@@ -2290,7 +4246,7 @@ int mrisFindAllOverlappingFaces(MRI_SURFACE *mris, MHT *mht, int fno, int *flist
   return (total_found);
 }
 
-int MRIScountNegativeFaces(MRI_SURFACE *mris)
+int MRIScountNegativeFaces(MRIS *mris)
 {
   int fno, neg;
   FACE *face;
@@ -2311,7 +4267,7 @@ int MRIScountNegativeFaces(MRI_SURFACE *mris)
   Description
   Scale a surface anisotropically.
   ------------------------------------------------------*/
-int MRISprintTessellationStats(MRI_SURFACE *mris, FILE *fp)
+int MRISprintTessellationStats(MRIS *mris, FILE *fp)
 {
   double mean, dsigma, dmin, dmax;
   int vno = 0, vno2 = 0;
@@ -2356,7 +4312,7 @@ int MRISprintTessellationStats(MRI_SURFACE *mris, FILE *fp)
   Description
   ------------------------------------------------------*/
 // get the vertex number
-int MRIStalairachToVertex(MRI_SURFACE *mris, double xt, double yt, double zt)
+int MRIStalairachToVertex(MRIS *mris, double xt, double yt, double zt)
 {
   int vno;
   double xw, yw, zw;
@@ -2378,7 +4334,7 @@ int MRIStalairachToVertex(MRI_SURFACE *mris, double xt, double yt, double zt)
 
   Description
   ------------------------------------------------------*/
-int MRIScanonicalToVertex(MRI_SURFACE *mris, double phi, double theta)
+int MRIScanonicalToVertex(MRIS *mris, double phi, double theta)
 {
   int vno;
   double xw, yw, zw;
@@ -2395,7 +4351,7 @@ int MRIScanonicalToVertex(MRI_SURFACE *mris, double phi, double theta)
 
   Description
   ------------------------------------------------------*/
-float MRISdistanceToSurface(MRI_SURFACE *mris, MHT *mht, float x0, float y0, float z0, float nx, float ny, float nz)
+float MRISdistanceToSurface(MRIS *mris, MHT *mht, float x0, float y0, float z0, float nx, float ny, float nz)
 {
   double dist, len;
 
@@ -2443,7 +4399,7 @@ float MRISdistanceToSurface(MRI_SURFACE *mris, MHT *mht, float x0, float y0, flo
 
   Description
   ------------------------------------------------------*/
-double MRISrmsTPHeight(MRI_SURFACE *mris)
+double MRISrmsTPHeight(MRIS *mris)
 {
   int vno, i, total_nbrs;
   double avg_height, dot, nx, ny, nz, x, y, z, d;
@@ -2498,7 +4454,7 @@ double MRISrmsTPHeight(MRI_SURFACE *mris)
   return (sqrt(avg_height / (double)total_nbrs));
 }
 /*!
-  \fn double mrisRmsValError(MRI_SURFACE *mris, MRI *mri)
+  \fn double mrisRmsValError(MRIS *mris, MRI *mri)
   \brief Samples mri at each vertex and computes the difference
   between the sample and v->val. The diff is squared and summed (SSE);
   that is then divided by the number of vertices hit and sqrt taken to
@@ -2508,7 +4464,7 @@ double MRISrmsTPHeight(MRI_SURFACE *mris)
   value of the sampled MRI value and v->val2bak takes the value
   of the error = sampled - target (ie, v->val2bak - v->val). #RMS
 */
-double mrisRmsValError(MRI_SURFACE *mris, MRI *mri)
+double mrisRmsValError(MRIS *mris, MRI *mri)
 {
   if (debugNonDeterminism) {
     mris_print_hash(stdout, mris, "mrisRmsValError", "\n");
@@ -2547,7 +4503,7 @@ double mrisRmsValError(MRI_SURFACE *mris, MRI *mri)
 
   Description
   ------------------------------------------------------*/
-double MRIScomputeAnalyticDistanceError(MRI_SURFACE *mris, int which, FILE *fp)
+double MRIScomputeAnalyticDistanceError(MRIS *mris, int which, FILE *fp)
 {
   int vno, n, vtotal, ndists;
   int const * pv;
@@ -2673,7 +4629,7 @@ double MRIScomputeAnalyticDistanceError(MRI_SURFACE *mris, int which, FILE *fp)
 
   Description
   ------------------------------------------------------*/
-double MRISstoreAnalyticDistances(MRI_SURFACE *mris, int which)
+double MRISstoreAnalyticDistances(MRIS *mris, int which)
 {
   int vno, n, vtotal;
   int const * pv;
@@ -2745,7 +4701,7 @@ double MRISstoreAnalyticDistances(MRI_SURFACE *mris, int which)
 
   Description
   ------------------------------------------------------*/
-int MRISdisturbOriginalDistances(MRI_SURFACE *mris, double max_pct)
+int MRISdisturbOriginalDistances(MRIS *mris, double max_pct)
 {
   int vno, n;
 
@@ -2772,7 +4728,7 @@ int MRISdisturbOriginalDistances(MRI_SURFACE *mris, double max_pct)
 */
 
 #include "MARS_DT_Boundary.h"
-int MRISdistanceTransform(MRI_SURFACE *mris, LABEL *area, int mode)
+int MRISdistanceTransform(MRIS *mris, LABEL *area, int mode)
 {
   int *vertices, *vertNbrs, vno, max_nbrs, j, index;
   double *cost, *vertDists;
@@ -2891,7 +4847,7 @@ int MRISdistanceTransform(MRI_SURFACE *mris, LABEL *area, int mode)
   assumes v->val has the distances for label1 and v->val2 has the distances
   for label 2, computed using MRISdistanceTransform
 */
-double MRIScomputeHausdorffDistance(MRI_SURFACE *mris, int mode)
+double MRIScomputeHausdorffDistance(MRIS *mris, int mode)
 {
   double hdist = 0, d, dist;
   int vno, n, num;
@@ -2962,7 +4918,7 @@ double MRIScomputeHausdorffDistance(MRI_SURFACE *mris, int mode)
   Compute the ratio of the two principal curvatures and
   store it in the vertex->curv variable.
   ------------------------------------------------------*/
-int MRISuseCurvatureRatio(MRI_SURFACE *mris)
+int MRISuseCurvatureRatio(MRIS *mris)
 {
   int vno;
   VERTEX *v;
@@ -3037,7 +4993,7 @@ int MRISuseCurvatureRatio(MRI_SURFACE *mris)
   Compute the contrast of the two principal curvatures and
   store it in the vertex->curv variable.
   ------------------------------------------------------*/
-int MRISuseCurvatureContrast(MRI_SURFACE *mris)
+int MRISuseCurvatureContrast(MRIS *mris)
 {
   int vno;
   VERTEX *v;
@@ -3097,7 +5053,7 @@ int MRISuseCurvatureContrast(MRI_SURFACE *mris)
 #include "mrishash.h"
 
 #define MAX_DIST 10
-int MRISmeasureDistanceBetweenSurfaces(MRI_SURFACE *mris, MRI_SURFACE *mris2, int signed_dist)
+int MRISmeasureDistanceBetweenSurfaces(MRIS *mris, MRIS *mris2, int signed_dist)
 {
   int vno;
   VERTEX *v1, *v2;
@@ -3668,7 +5624,7 @@ void computeDefectFaceNormal(MRIS const * const mris, int const fno)
 }
 
 
-int MRIScomputeSurfaceNormals(MRI_SURFACE *mris, int which, int navgs)
+int MRIScomputeSurfaceNormals(MRIS *mris, int which, int navgs)
 {
   int vno, n, i;
   double nx = 0.0, ny = 0.0, nz = 0.0, norm = 0.0;
@@ -3815,7 +5771,7 @@ int MRIScomputeSurfaceNormals(MRI_SURFACE *mris, int which, int navgs)
 }
 
 
-int MRISsetCurvature(MRI_SURFACE *mris, float val)
+int MRISsetCurvature(MRIS *mris, float val)
 {
   int vno;
   VERTEX *v;
@@ -3830,7 +5786,7 @@ int MRISsetCurvature(MRI_SURFACE *mris, float val)
   return (NO_ERROR);
 }
 
-int MRIStrinarizeCurvature(MRI_SURFACE *mris, float trinarize_thresh)
+int MRIStrinarizeCurvature(MRIS *mris, float trinarize_thresh)
 {
   int vno;
   VERTEX *v;
@@ -3854,7 +5810,7 @@ int MRIStrinarizeCurvature(MRI_SURFACE *mris, float trinarize_thresh)
 }
 
 
-int MRISscaleCurvature(MRI_SURFACE *apmris, float af_scale)
+int MRISscaleCurvature(MRIS *apmris, float af_scale)
 {
   //
   // POSTCONDITIONS
@@ -3891,7 +5847,7 @@ int MRISscaleCurvature(MRI_SURFACE *apmris, float af_scale)
 }
 
 
-int MRISbinarizeCurvature(MRI_SURFACE *mris, float thresh, float low, float high, int use_abs)
+int MRISbinarizeCurvature(MRIS *mris, float thresh, float low, float high, int use_abs)
 {
   int vno;
   VERTEX *v;
@@ -3913,7 +5869,7 @@ int MRISbinarizeCurvature(MRI_SURFACE *mris, float thresh, float low, float high
 
   return (NO_ERROR);
 }
-int MRISthresholdCurvature(MRI_SURFACE *mris, float thresh, int use_abs)
+int MRISthresholdCurvature(MRIS *mris, float thresh, int use_abs)
 {
   int vno;
   VERTEX *v;
@@ -3982,7 +5938,7 @@ void MRISnormalizeField(MRIS *mris, int distance_field, int which_norm)
 
 
 // remove outliers in the v->curv field
-int MRIShistoThresholdCurvature(MRI_SURFACE *mris, float thresh_pct)
+int MRIShistoThresholdCurvature(MRIS *mris, float thresh_pct)
 {
   double min_curv, max_curv, curv_scale, total, thresh;
   int bin, zbin, nthresh = 0, vno;
@@ -4103,7 +6059,7 @@ int MRIShistoThresholdCurvature(MRI_SURFACE *mris, float thresh_pct)
   the pial surface.
   ------------------------------------------------------*/
 
-int MRISfindClosestOrigVertices(MRI_SURFACE *mris, int nbhd_size)
+int MRISfindClosestOrigVertices(MRIS *mris, int nbhd_size)
 {
   int vno, n, vlist[100000], vtotal, ns, i, vnum, nbr_count[100], min_n, min_vno;
   float dx, dy, dz, dist, min_dist, nx, ny, nz, dot;
@@ -4195,7 +6151,7 @@ int MRISfindClosestOrigVertices(MRI_SURFACE *mris, int nbhd_size)
   find the closest pial vertex and use it's spherical coords to initialize
   the thickness minimization. Put the v->c[xyz] coords of the nearest pial vertex into v->[xyz] of each vertex.
 */
-int MRISfindClosestPialVerticesCanonicalCoords(MRI_SURFACE *mris, int nbhd_size)
+int MRISfindClosestPialVerticesCanonicalCoords(MRIS *mris, int nbhd_size)
 {
   int vno, n, vlist[100000], vtotal, ns, i, vnum, nbr_count[100], min_n, min_vno;
   float dx, dy, dz, dist, min_dist, nx, ny, nz, dot;
@@ -4287,7 +6243,7 @@ int MRISfindClosestPialVerticesCanonicalCoords(MRI_SURFACE *mris, int nbhd_size)
 }
 
 
-int MRISmeasureCorticalThickness(MRI_SURFACE *mris, int nbhd_size, float max_thick)
+int MRISmeasureCorticalThickness(MRIS *mris, int nbhd_size, float max_thick)
 {
   int vno, n, vlist[100000], vtotal, ns, i, vnum, nbr_count[100], min_n, nwg_bad, ngw_bad;
   float dx, dy, dz, dist, min_dist, nx, ny, nz, dot;
@@ -5130,7 +7086,7 @@ static float mrisTriangleArea(MRIS *mris, int fac, int n)
 
   Description
   ------------------------------------------------------*/
-int MRIScountNegativeTriangles(MRI_SURFACE *mris)
+int MRIScountNegativeTriangles(MRIS *mris)
 {
   int fno, negative;
   FACE *face;
@@ -5150,7 +7106,7 @@ int MRIScountNegativeTriangles(MRI_SURFACE *mris)
 }
 
 
-int mrisComputeSurfaceDimensions(MRI_SURFACE *mris)
+int mrisComputeSurfaceDimensions(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -5209,7 +7165,7 @@ int mrisComputeSurfaceDimensions(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int mrisClearDistances(MRI_SURFACE *mris)
+int mrisClearDistances(MRIS *mris)
 {
   int vno, nvertices;
   VERTEX *v;
@@ -5229,7 +7185,7 @@ int mrisClearDistances(MRI_SURFACE *mris)
 /*-----------------------------------------------------------------
   Calculate distances between each vertex and each of its neighbors.
   ----------------------------------------------------------------*/
-int mrisComputeVertexDistances(MRI_SURFACE *mris)
+int mrisComputeVertexDistances(MRIS *mris)
 {
   int vno;
 
@@ -5324,7 +7280,7 @@ int mrisComputeVertexDistances(MRI_SURFACE *mris)
 }
 
 
-int mrisComputeOriginalVertexDistances(MRI_SURFACE *mris)
+int mrisComputeOriginalVertexDistances(MRIS *mris)
 {
   int vno, n, vtotal, *pv;
   float d, xd, yd, zd, circumference = 0.0f, angle;
@@ -5504,7 +7460,7 @@ void mrisSetAvgInterVertexDist(MRIS *mris, double to)
   Description
   same as  below, but tracks odx,ody,odz fields
   ------------------------------------------------------*/
-int mrisTrackTotalDistanceNew(MRI_SURFACE *mris)
+int mrisTrackTotalDistanceNew(MRIS *mris)
 {
   int vno;
   VERTEX *v;
@@ -5527,7 +7483,7 @@ int mrisTrackTotalDistanceNew(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int mrisTrackTotalDistance(MRI_SURFACE *mris)
+int mrisTrackTotalDistance(MRIS *mris)
 {
   int vno;
   VERTEX *v;
@@ -5544,7 +7500,7 @@ int mrisTrackTotalDistance(MRI_SURFACE *mris)
   return (NO_ERROR);
 }
 
-int MRIScomputeAllDistances(MRI_SURFACE *mris)
+int MRIScomputeAllDistances(MRIS *mris)
 {
   int vno, done = 0, nvalid;
   LABEL *area;
@@ -5637,8 +7593,8 @@ int MRIScomputeAllDistances(MRI_SURFACE *mris)
 
 
 #ifdef BEVIN_MRISCOMPUTENORMALS_REPRODUCIBLE
-static int MRIScomputeNormals_old(MRI_SURFACE *mris);
-static int MRIScomputeNormals_new(MRI_SURFACE *mris);
+static int MRIScomputeNormals_old(MRIS *mris);
+static int MRIScomputeNormals_new(MRIS *mris);
 #endif
 
 #ifdef BEVIN_MRISCOMPUTENORMALS_CHECK
@@ -5657,7 +7613,7 @@ static int MRIScomputeNormals_new(MRI_SURFACE *mris);
     
     static void MRIScomputeNormals_Snapshot_init(
         MRIScomputeNormals_Snapshot* lhs,
-        MRI_SURFACE*                 mris) {
+        MRIS*                 mris) {
 	int CC = mris->nfaces    * sizeof(FaceNormCacheEntry);  memcpy(lhs->faceNormCacheEntries = (FaceNormCacheEntry*  )malloc(CC), mris->faceNormCacheEntries, CC); 
         int FC = mris->nfaces    * sizeof(FACE);   		memcpy(lhs->faces                = (FACE*  )              malloc(FC), mris->faces,                FC); 
         int VC = mris->nvertices * sizeof(VERTEX); 		memcpy(lhs->vertices             = (VERTEX*)              malloc(VC), mris->vertices,             VC);
@@ -5665,7 +7621,7 @@ static int MRIScomputeNormals_new(MRI_SURFACE *mris);
 
     static void MRIScomputeNormals_Snapshot_undo(
         MRIScomputeNormals_Snapshot* lhs,
-        MRI_SURFACE*                 mris) {
+        MRIS*                 mris) {
 	int CC = mris->nfaces    * sizeof(FaceNormCacheEntry);  memcpy(mris->faceNormCacheEntries, lhs->faceNormCacheEntries, CC);
         int FC = mris->nfaces    * sizeof(FACE);   		memcpy(mris->faces,                lhs->faces,                FC); 
         int VC = mris->nvertices * sizeof(VERTEX); 		memcpy(mris->vertices,             lhs->vertices,             VC); 
@@ -5691,7 +7647,7 @@ static int MRIScomputeNormals_new(MRI_SURFACE *mris);
     static void MRIScomputeNormals_Snapshot_compare(
         MRIScomputeNormals_Snapshot* lhs,
         MRIScomputeNormals_Snapshot* rhs,
-        MRI_SURFACE*                 mris) {
+        MRIS*                 mris) {
         
         int count = 0;
         int i;
@@ -5731,7 +7687,7 @@ static int MRIScomputeNormals_new(MRI_SURFACE *mris);
 #endif
 
 
-int MRIScomputeNormals(MRI_SURFACE *mris)
+int MRIScomputeNormals(MRIS *mris)
 #ifdef BEVIN_MRISCOMPUTENORMALS_REPRODUCIBLE
 {
 #ifdef BEVIN_MRISCOMPUTENORMALS_CHECK
@@ -5758,7 +7714,7 @@ int MRIScomputeNormals(MRI_SURFACE *mris)
     return result;
 }
 
-static int MRIScomputeNormals_old(MRI_SURFACE *mris)
+static int MRIScomputeNormals_old(MRIS *mris)
 #endif
 {
   // This is the old code.  It has several parallelization problems, as noted in the code
@@ -5922,7 +7878,7 @@ static int int_compare(const void* lhs_ptr, const void* rhs_ptr) {
    return lhs - rhs;
 }
 
-static int MRIScomputeNormals_new(MRI_SURFACE *mris)
+static int MRIScomputeNormals_new(MRIS *mris)
 {
   static const double RAN = 0.001; /* one thousandth of a millimeter */
 
@@ -6224,10 +8180,10 @@ static int MRIScomputeNormals_new(MRI_SURFACE *mris)
 
 
 #ifdef BEVIN_MRISCOMPUTETRIANGLEPROPERTIES_REPRODUCIBLE
-static int MRIScomputeTriangleProperties_old(MRI_SURFACE *mris, bool old_done);
-static int MRIScomputeTriangleProperties_new(MRI_SURFACE *mris, bool old_done);
+static int MRIScomputeTriangleProperties_old(MRIS *mris, bool old_done);
+static int MRIScomputeTriangleProperties_new(MRIS *mris, bool old_done);
 
-int MRIScomputeTriangleProperties(MRI_SURFACE *mris)
+int MRIScomputeTriangleProperties(MRIS *mris)
 {
 
     static const bool do_old = 
@@ -6252,7 +8208,7 @@ int MRIScomputeTriangleProperties(MRI_SURFACE *mris)
     return result_new;
 }
 
-static int MRIScomputeTriangleProperties_old(MRI_SURFACE *mris, bool old_done)
+static int MRIScomputeTriangleProperties_old(MRIS *mris, bool old_done)
 
 #endif
 
@@ -6406,7 +8362,7 @@ static int MRIScomputeTriangleProperties_old(MRI_SURFACE *mris, bool old_done)
 
 
 #ifdef BEVIN_MRISCOMPUTETRIANGLEPROPERTIES_REPRODUCIBLE
-static int MRIScomputeTriangleProperties_new(MRI_SURFACE *mris, bool old_done)
+static int MRIScomputeTriangleProperties_new(MRIS *mris, bool old_done)
 {
   int const max_threads = omp_get_max_threads();
   
@@ -6646,7 +8602,7 @@ static int MRIScomputeTriangleProperties_new(MRI_SURFACE *mris, bool old_done)
 
   Description
   ------------------------------------------------------*/
-double MRIStotalVariationDifference(MRI_SURFACE *mris)
+double MRIStotalVariationDifference(MRIS *mris)
 {
   int vno, nvertices;
   VERTEX *v;
@@ -6673,7 +8629,7 @@ double MRIStotalVariationDifference(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-double MRIStotalVariation(MRI_SURFACE *mris)
+double MRIStotalVariation(MRIS *mris)
 {
   int vno, nvertices;
   VERTEX *v;
@@ -6703,7 +8659,7 @@ double MRIStotalVariation(MRI_SURFACE *mris)
   If so, it should not be added to the tessellation.
   ------------------------------------------------------*/
 #if SPHERE_INTERSECTION
-int containsAnotherVertexOnSphere(MRI_SURFACE *mris, int vno0, int vno1, int vno2, int mode)
+int containsAnotherVertexOnSphere(MRIS *mris, int vno0, int vno1, int vno2, int mode)
 {
   int i, n, nvertices, vertices[1000];
   double sign[3], normal[3][3], orgn[3][3], tangent[3], test;
@@ -6810,7 +8766,7 @@ int containsAnotherVertexOnSphere(MRI_SURFACE *mris, int vno0, int vno1, int vno
 }
 #else
 int containsAnotherVertex(
-    MRI_SURFACE *mris, int vno0, int vno1, int vno2, double e0[3], double e1[3], double origin[3])
+    MRIS *mris, int vno0, int vno1, int vno2, double e0[3], double e1[3], double origin[3])
 #if 1
 {
   int n, vno, i, n1;
@@ -7084,7 +9040,7 @@ int containsAnotherVertex(
 
   Description
   ------------------------------------------------------*/
-double MRIScurvatureError(MRI_SURFACE *mris, double Kd)
+double MRIScurvatureError(MRIS *mris, double Kd)
 {
   int vno, nvertices, n;
   VERTEX *v;
@@ -7117,7 +9073,7 @@ double MRIScurvatureError(MRI_SURFACE *mris, double Kd)
   Description
   ------------------------------------------------------*/
 int MRISsurfaceRASToTalairachVoxel(
-    MRI_SURFACE *mris, MRI *mri, double xw, double yw, double zw, double *pxv, double *pyv, double *pzv)
+    MRIS *mris, MRI *mri, double xw, double yw, double zw, double *pxv, double *pyv, double *pzv)
 {
   //   conformed -->  SRAS
   //       |            |
@@ -7179,7 +9135,7 @@ int MRISsurfaceRASToTalairachVoxel(
 
   Description
   ------------------------------------------------------*/
-int MRISvertexToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pxv, double *pyv, double *pzv)
+int MRISvertexToVoxel(MRIS *mris, VERTEX *v, MRI *mri, double *pxv, double *pyv, double *pzv)
 {
   double xw, yw, zw;
 
@@ -7190,7 +9146,7 @@ int MRISvertexToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pxv, doubl
   return (NO_ERROR);
 }
 
-int MRISvertexNormalToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pnx, double *pny, double *pnz)
+int MRISvertexNormalToVoxel(MRIS *mris, VERTEX *v, MRI *mri, double *pnx, double *pny, double *pnz)
 {
   double xw, yw, zw;
   double xv0, yv0, zv0, xv1, yv1, zv1;
@@ -7208,7 +9164,7 @@ int MRISvertexNormalToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pnx,
   *pnz = zv1 - zv0;
   return (NO_ERROR);
 }
-int MRISvertexNormalToVoxelScaled(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pnx, double *pny, double *pnz)
+int MRISvertexNormalToVoxelScaled(MRIS *mris, VERTEX *v, MRI *mri, double *pnx, double *pny, double *pnz)
 {
   double nx, ny, nz, len;
 
@@ -7227,7 +9183,7 @@ int MRISvertexNormalToVoxelScaled(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double
 
   Description
   ------------------------------------------------------*/
-int MRISvertexCoordToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, int coords, double *pxv, double *pyv, double *pzv)
+int MRISvertexCoordToVoxel(MRIS *mris, VERTEX *v, MRI *mri, int coords, double *pxv, double *pyv, double *pzv)
 {
   double xw = 0., yw = 0., zw = 0.;
 
@@ -7242,7 +9198,7 @@ int MRISvertexCoordToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, int coords, d
 
   Description
   ------------------------------------------------------*/
-int MRISorigVertexToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pxv, double *pyv, double *pzv)
+int MRISorigVertexToVoxel(MRIS *mris, VERTEX *v, MRI *mri, double *pxv, double *pyv, double *pzv)
 {
   double xw, yw, zw;
 
@@ -7259,7 +9215,7 @@ int MRISorigVertexToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pxv, d
 
   Description
   ------------------------------------------------------*/
-int MRISwhiteVertexToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pxv, double *pyv, double *pzv)
+int MRISwhiteVertexToVoxel(MRIS *mris, VERTEX *v, MRI *mri, double *pxv, double *pyv, double *pzv)
 {
   double xw, yw, zw;
 
@@ -7275,7 +9231,7 @@ int MRISwhiteVertexToVoxel(MRI_SURFACE *mris, VERTEX *v, MRI *mri, double *pxv, 
   indicate the correspondence between white and pial, and the canonical spherical ones v->c[xyz] have the ?h.sphere in
   them.
 */
-int MRISmeasureThicknessFromCorrespondence(MRI_SURFACE *mris, MHT *mht, float max_thick)
+int MRISmeasureThicknessFromCorrespondence(MRIS *mris, MHT *mht, float max_thick)
 {
   int vno, alloced;
   VERTEX *v;
@@ -7309,7 +9265,7 @@ int MRISmeasureThicknessFromCorrespondence(MRI_SURFACE *mris, MHT *mht, float ma
 }
 
 
-int mrisComputeNormalDotDistribution(MRI_SURFACE *mris, HISTOGRAM *h_dot)
+int mrisComputeNormalDotDistribution(MRIS *mris, HISTOGRAM *h_dot)
 {
   int vno, bin, n, num;
 
@@ -7411,7 +9367,7 @@ int mrisComputeNormalDotDistribution(MRI_SURFACE *mris, HISTOGRAM *h_dot)
   MRIScomputeMetricProperties(mris);
   return (NO_ERROR);
 }
-int mrisComputePrincipalCurvatureDistributions(MRI_SURFACE *mris,
+int mrisComputePrincipalCurvatureDistributions(MRIS *mris,
                                                       HISTOGRAM *h_k1,
                                                       HISTOGRAM *h_k2,
                                                       MRI *mri_k1_k2)
@@ -7547,7 +9503,7 @@ int mrisComputePrincipalCurvatureDistributions(MRI_SURFACE *mris,
 
 //=====================================================================================================================
 static int get_face_axes(
-    MRI_SURFACE *mris, FACE *face, float *pe1x, float *pe1y, float *pe1z, float *pe2x, float *pe2y, float *pe2z)
+    MRIS *mris, FACE *face, float *pe1x, float *pe1y, float *pe1z, float *pe2x, float *pe2y, float *pe2z)
 {
   double V0[3], V1[3], V2[3], e1[3], e2[3], tmp[3], norm, dot;
 
@@ -7619,7 +9575,7 @@ static int get_face_axes(
 }
 
 
-static double ashburnerTriangleEnergy(MRI_SURFACE *mris, int fno, double lambda)
+static double ashburnerTriangleEnergy(MRIS *mris, int fno, double lambda)
 {
   static MATRIX *_m_x[_MAX_FS_THREADS], *_m_x_inv[_MAX_FS_THREADS], *_m_y[_MAX_FS_THREADS], *_m_J[_MAX_FS_THREADS],
       *_m_U[_MAX_FS_THREADS], *_m_V[_MAX_FS_THREADS], *_m_m[_MAX_FS_THREADS], *_m_evectors[_MAX_FS_THREADS];
@@ -7760,7 +9716,7 @@ static double ashburnerTriangleEnergy(MRI_SURFACE *mris, int fno, double lambda)
 
 
 float mrisSampleAshburnerTriangleEnergy(
-    MRI_SURFACE * const mris, int const vno, INTEGRATION_PARMS * const parms, float cx, float cy, float cz)
+    MRIS * const mris, int const vno, INTEGRATION_PARMS * const parms, float cx, float cy, float cz)
 {
   VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
   VERTEX                * const v  = &mris->vertices         [vno];
@@ -7797,7 +9753,7 @@ float mrisSampleAshburnerTriangleEnergy(
 
   Description
   ------------------------------------------------------*/
-double mrisComputeThicknessSmoothnessEnergy(MRI_SURFACE *mris, double l_tsmooth, INTEGRATION_PARMS *parms)
+double mrisComputeThicknessSmoothnessEnergy(MRIS *mris, double l_tsmooth, INTEGRATION_PARMS *parms)
 {
   int vno, n;
   double sse_tsmooth, v_sse, dn, dx, dy, dz, d0;
@@ -7835,7 +9791,7 @@ double mrisComputeThicknessSmoothnessEnergy(MRI_SURFACE *mris, double l_tsmooth,
 }
 
 float mrisSampleMinimizationEnergy(
-    MRI_SURFACE *mris, VERTEX *v, INTEGRATION_PARMS *parms, float cx, float cy, float cz)
+    MRIS *mris, VERTEX *v, INTEGRATION_PARMS *parms, float cx, float cy, float cz)
 {
   float xw, yw, zw, dx, dy, dz, thick_sq, xp, yp, zp;
 
@@ -7851,7 +9807,7 @@ float mrisSampleMinimizationEnergy(
   return (thick_sq);
 }
 
-float mrisSampleParallelEnergyAtVertex(MRI_SURFACE *mris, int const vno, INTEGRATION_PARMS *parms)
+float mrisSampleParallelEnergyAtVertex(MRIS *mris, int const vno, INTEGRATION_PARMS *parms)
 {
   VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
   VERTEX                * const v  = &mris->vertices         [vno];
@@ -7927,7 +9883,7 @@ float mrisSampleParallelEnergyAtVertex(MRI_SURFACE *mris, int const vno, INTEGRA
 }
 
 float mrisSampleParallelEnergy(
-    MRI_SURFACE *mris, int const vno, INTEGRATION_PARMS *parms, float cx, float cy, float cz)
+    MRIS *mris, int const vno, INTEGRATION_PARMS *parms, float cx, float cy, float cz)
 {
   VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
   VERTEX                * const v  = &mris->vertices         [vno];
@@ -7976,7 +9932,7 @@ float mrisSampleParallelEnergy(
 }
 
 float mrisSampleNormalEnergy(
-    MRI_SURFACE *mris, VERTEX *v, INTEGRATION_PARMS *parms, float cx, float cy, float cz)
+    MRIS *mris, VERTEX *v, INTEGRATION_PARMS *parms, float cx, float cy, float cz)
 {
   float dx, dy, dz, len, xw, yw, zw, xp, yp, zp, pnx, pny, pnz;
   double sse;
@@ -8021,7 +9977,7 @@ float mrisSampleNormalEnergy(
 }
 
 float mrisSampleSpringEnergy(
-    MRI_SURFACE *mris, int const vno, float cx, float cy, float cz, INTEGRATION_PARMS *parms)
+    MRIS *mris, int const vno, float cx, float cy, float cz, INTEGRATION_PARMS *parms)
 {
   VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
   VERTEX          const * const v  = &mris->vertices         [vno];
@@ -8073,7 +10029,7 @@ float mrisSampleSpringEnergy(
 
   Description
   ------------------------------------------------------*/
-double mrisComputeThicknessMinimizationEnergy(MRI_SURFACE *mris, double l_thick_min, INTEGRATION_PARMS *parms)
+double mrisComputeThicknessMinimizationEnergy(MRIS *mris, double l_thick_min, INTEGRATION_PARMS *parms)
 {
   int vno;
   double sse_tmin;
@@ -8131,7 +10087,7 @@ double mrisComputeThicknessMinimizationEnergy(MRI_SURFACE *mris, double l_thick_
 
   Description
   ------------------------------------------------------*/
-int mrisComputeCurvatureMinMax(MRI_SURFACE *mris)
+int mrisComputeCurvatureMinMax(MRIS *mris)
 {
   int vno, found = 0;
   VERTEX *v;
@@ -8164,7 +10120,7 @@ int mrisComputeCurvatureMinMax(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISuseCurvatureDifference(MRI_SURFACE *mris)
+int MRISuseCurvatureDifference(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -8197,7 +10153,7 @@ int MRISuseCurvatureDifference(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISuseCurvatureMax(MRI_SURFACE *mris)
+int MRISuseCurvatureMax(MRIS *mris)
 {
   int vno;
   VERTEX *v;
@@ -8240,7 +10196,7 @@ int MRISuseCurvatureMax(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISuseCurvatureMin(MRI_SURFACE *mris)
+int MRISuseCurvatureMin(MRIS *mris)
 {
   int vno;
   VERTEX *v;
@@ -8282,7 +10238,7 @@ int MRISuseCurvatureMin(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISuseCurvatureStretch(MRI_SURFACE *mris)
+int MRISuseCurvatureStretch(MRIS *mris)
 {
   int vno, n;
   float kmin, kmax, dist, dist_orig, curv, dist_scale, max_stretch, stretch;
@@ -8329,7 +10285,7 @@ int MRISuseCurvatureStretch(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISuseNegCurvature(MRI_SURFACE *mris)
+int MRISuseNegCurvature(MRIS *mris)
 {
   int vno, fno;
   FACE *f;
@@ -8362,7 +10318,7 @@ int MRISuseNegCurvature(MRI_SURFACE *mris)
   Description
   ------------------------------------------------------*/
 static double big_sse = 10.0;
-double mrisComputeThicknessNormalEnergy(MRI_SURFACE *mris, double l_thick_normal, INTEGRATION_PARMS *parms)
+double mrisComputeThicknessNormalEnergy(MRIS *mris, double l_thick_normal, INTEGRATION_PARMS *parms)
 {
   int vno;
   double sse_tnormal;
@@ -8443,7 +10399,7 @@ double mrisComputeThicknessNormalEnergy(MRI_SURFACE *mris, double l_thick_normal
 
   Description
   ------------------------------------------------------*/
-double mrisComputeThicknessSpringEnergy(MRI_SURFACE *mris, double l_thick_spring, INTEGRATION_PARMS *parms)
+double mrisComputeThicknessSpringEnergy(MRIS *mris, double l_thick_spring, INTEGRATION_PARMS *parms)
 {
   int vno;
   double sse_spring, sse;
@@ -8652,7 +10608,7 @@ static void MRISaverageGradients_write(
   fprintf(stdout, "%s:%d finished writing %s\n",__FILE__, __LINE__, filename);
 }
 
-int MRISaverageGradients(MRI_SURFACE *mris, int num_avgs)
+int MRISaverageGradients(MRIS *mris, int num_avgs)
 {
   int i, vno;
   float sigma;
@@ -9075,7 +11031,7 @@ int MRISaverageGradients(MRI_SURFACE *mris, int num_avgs)
   of MRISaverageGradients(). Should be about 40% faster. It also gives
   identical results as can be verified with MRISaverageGradientsFastCheck().
   ------------------------------------------------------*/
-int MRISaverageGradientsFast(MRI_SURFACE *mris, int num_avgs)
+int MRISaverageGradientsFast(MRIS *mris, int num_avgs)
 {
   int vno, vnb, *pnb, num, nthavg;
 
@@ -9299,6 +11255,153 @@ int MRISaverageGradientsFastCheck(int num_avgs)
 }
 
 
+int mrisAverageSignedGradients(MRIS *mris, int num_avgs)
+{
+  int i, vno;
+  float sigma;
+
+  MRI_SP *mrisp, *mrisp_blur;
+
+  if (num_avgs <= 0) {
+    return (NO_ERROR);
+  }
+
+  if (Gdiag_no >= 0) {
+    VERTEX const * const v = &mris->vertices[Gdiag_no];
+    fprintf(stdout, "before averaging dot = %2.2f ", v->dx * v->nx + v->dy * v->ny + v->dz * v->nz);
+
+    {
+      int vno;
+      for (vno = 0; vno < mris->nvertices; vno++) {
+        VERTEX const * const vn = &mris->vertices[vno];
+        if (vn->x > 1 && vn->dx > .1) DiagBreak();
+      }
+    }
+    {
+      VERTEX const *vmax = NULL;
+      double mx = 0;
+      int vno;
+      for (vno = 0; vno < mris->nvertices; vno++) {
+        VERTEX const * const vn = &mris->vertices[vno];
+        if (vn->x > 1 && vn->dx > .1) DiagBreak();
+        if (vn->dx > mx && vn->x > 1) {
+          mx = vn->dx;
+          vmax = vn;
+        }
+      }
+      if (mx > 0) DiagBreak();
+    }
+  }
+  if (0 && mris->status == MRIS_PARAMETERIZED_SPHERE) /* use convolution */
+  {
+    sigma = sqrt((float)num_avgs) / 4.0;
+    mrisp = MRISgradientToParameterization(mris, NULL, 1.0);
+    mrisp_blur = MRISPblur(mrisp, NULL, sigma, -1);
+    MRISgradientFromParameterization(mrisp_blur, mris);
+    MRISPfree(&mrisp);
+    MRISPfree(&mrisp_blur);
+  }
+  else
+    for (i = 0; i < num_avgs; i++) {
+      ROMP_PF_begin
+#ifdef HAVE_OPENMP
+      #pragma omp parallel for if_ROMP(experimental)
+#endif
+      for (vno = 0; vno < mris->nvertices; vno++) {
+        ROMP_PFLB_begin
+	
+        double dx, dy, dz, dot;
+        int vnum, num, vnb;
+
+        VERTEX_TOPOLOGY const * const vt = &mris->vertices_topology[vno];
+        VERTEX                * const v  = &mris->vertices         [vno];
+        if (v->ripflag) {
+          continue;
+        }
+        dx = v->dx;
+        dy = v->dy;
+        dz = v->dz;
+        int const * pnb = vt->v;
+        /*      vnum = v->v2num ? v->v2num : v->vnum ;*/
+        vnum = vt->vnum;
+        for (num = 0, vnb = 0; vnb < vnum; vnb++) {
+          VERTEX const * const vn = &mris->vertices[*pnb++]; /* neighboring vertex pointer */
+          if (vn->ripflag) {
+            continue;
+          }
+          dot = vn->dx * v->dx + vn->dy * v->dy + vn->dz * v->dz;
+          if (dot < 0) {
+            continue; /* pointing in opposite directions */
+          }
+
+          num++;
+          dx += vn->dx;
+          dy += vn->dy;
+          dz += vn->dz;
+#if 0
+          if (vno == Gdiag_no)
+          {
+            float dot ;
+            dot = vn->dx*v->dx + vn->dy*v->dy + vn->dz*v->dz ;
+            if (dot < 0)
+              fprintf(stdout,
+                      "vn %d: dot = %2.3f, dx = (%2.3f, %2.3f, %2.3f)\n",
+                      v->v[vnb], dot, vn->dx, vn->dy, vn->dz) ;
+          }
+#endif
+        }
+        num++;
+        v->tdx = dx / (float)num;
+        v->tdy = dy / (float)num;
+        v->tdz = dz / (float)num;
+	ROMP_PFLB_end
+      }
+      ROMP_PF_end
+      
+      ROMP_PF_begin
+#ifdef HAVE_OPENMP
+      #pragma omp parallel for if_ROMP(experimental)
+#endif
+      for (vno = 0; vno < mris->nvertices; vno++) {
+        ROMP_PFLB_begin
+	
+        VERTEX *v = &mris->vertices[vno];
+        if (v->ripflag) ROMP_PF_continue;
+
+        v->dx = v->tdx;
+        v->dy = v->tdy;
+        v->dz = v->tdz;
+	ROMP_PFLB_end
+      }
+      ROMP_PF_end
+    }
+  if (Gdiag_no >= 0) {
+    float dot;
+    VERTEX const * const v = &mris->vertices[Gdiag_no];
+    dot = v->nx * v->dx + v->ny * v->dy + v->nz * v->dz;
+    fprintf(stdout, " after dot = %2.2f (%2.3f, %2.3f, %2.3f)\n", dot, v->dx, v->dy, v->dz);
+    if (fabs(dot) > 50) {
+      DiagBreak();
+    }
+    {
+      VERTEX const *vmax = NULL;
+      double mx = 0;
+      int vno;
+      for (vno = 0; vno < mris->nvertices; vno++) {
+        VERTEX const * const vn = &mris->vertices[vno];
+        if (vn->x > 1 && vn->dx > .1) DiagBreak();
+        if (vn->dx > mx && vn->x > 1) {
+          mx = vn->dx;
+          vmax = vn;
+        }
+      }
+      if (mx > 0) DiagBreak();
+    }
+  }
+  return (NO_ERROR);
+}
+
+
 /*-----------------------------------------------------
   Parameters:
 
@@ -9306,7 +11409,7 @@ int MRISaverageGradientsFastCheck(int num_avgs)
 
   Description
   ------------------------------------------------------*/
-int MRISsmoothSurfaceNormals(MRI_SURFACE *mris, int navgs)
+int MRISsmoothSurfaceNormals(MRIS *mris, int navgs)
 {
   int i, vno, vnb, vnum;
   float nx, ny, nz, num, len;
@@ -9363,7 +11466,7 @@ int MRISsmoothSurfaceNormals(MRI_SURFACE *mris, int navgs)
 
   Description
   ------------------------------------------------------*/
-int mrisComputeBoundaryNormals(MRI_SURFACE *mris)
+int mrisComputeBoundaryNormals(MRIS *mris)
 {
 #if 0
   int      k,m,n;
@@ -9432,7 +11535,7 @@ int mrisComputeBoundaryNormals(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRIScomputeMeanCurvature(MRI_SURFACE *mris)
+int MRIScomputeMeanCurvature(MRIS *mris)
 {
   VECTOR *v_n, *v_e1, *v_e2, *v_i;
   int vno, i, N;
@@ -9505,12 +11608,12 @@ int MRIScomputeMeanCurvature(MRI_SURFACE *mris)
   return (NO_ERROR);
 }
 
-int MRIScomputeSecondFundamentalForm(MRI_SURFACE *mris)
+int MRIScomputeSecondFundamentalForm(MRIS *mris)
 {
   return (MRIScomputeSecondFundamentalFormThresholded(mris, -1));
 }
 
-int MRIScomputeSecondFundamentalFormThresholded(MRI_SURFACE *mris, double pct_thresh)
+int MRIScomputeSecondFundamentalFormThresholded(MRIS *mris, double pct_thresh)
 {
   double min_k1, min_k2, max_k1, max_k2, k1_scale, k2_scale, total, thresh, orig_rsq_thresh;
   int bin, zbin1, zbin2, nthresh = 0;
@@ -9962,7 +12065,7 @@ int MRIScomputeSecondFundamentalFormThresholded(MRI_SURFACE *mris, double pct_th
   return (NO_ERROR);
 }
 
-int MRIScomputeSecondFundamentalFormAtVertex(MRI_SURFACE *mris, int vno, int *vertices, int vnum)
+int MRIScomputeSecondFundamentalFormAtVertex(MRIS *mris, int vno, int *vertices, int vnum)
 {
   int i, n, nbad = 0;
   VERTEX *vertex, *vnb;
@@ -10263,7 +12366,7 @@ void updateDistanceEltFromSignArgAndSquareNoLockNeeded(
 
   Description
   ------------------------------------------------------*/
-int MRISuseAreaErrors(MRI_SURFACE *mris)
+int MRISuseAreaErrors(MRIS *mris)
 {
   int vno, fi, n;
   float area, orig_area;
@@ -10299,7 +12402,7 @@ int MRISuseAreaErrors(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISuseGaussianCurvature(MRI_SURFACE *mris)
+int MRISuseGaussianCurvature(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -10317,10 +12420,10 @@ int MRISuseGaussianCurvature(MRI_SURFACE *mris)
   return (NO_ERROR);
 }
 /*-----------------------------------------------------
-  int MRISuseMeanCurvature(MRI_SURFACE *mris)
+  int MRISuseMeanCurvature(MRIS *mris)
   Set vertex->curv = vertex->H ;
   ------------------------------------------------------*/
-int MRISuseMeanCurvature(MRI_SURFACE *mris)
+int MRISuseMeanCurvature(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -10344,7 +12447,7 @@ int MRISuseMeanCurvature(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISusePrincipalCurvature(MRI_SURFACE *mris)
+int MRISusePrincipalCurvature(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -10362,7 +12465,7 @@ int MRISusePrincipalCurvature(MRI_SURFACE *mris)
   return (NO_ERROR);
 }
 
-int MRISuseK1Curvature(MRI_SURFACE *mris)
+int MRISuseK1Curvature(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -10389,7 +12492,7 @@ int MRISuseK1Curvature(MRI_SURFACE *mris)
   return (NO_ERROR);
 }
 
-int MRISuseK2Curvature(MRI_SURFACE *mris)
+int MRISuseK2Curvature(MRIS *mris)
 {
   int vno;
   VERTEX *vertex;
@@ -10416,7 +12519,7 @@ int MRISuseK2Curvature(MRI_SURFACE *mris)
   return (NO_ERROR);
 }
 
-int MRISusePrincipalCurvatureFunction(MRI_SURFACE *pmris, float (*f)(float k1, float k2))
+int MRISusePrincipalCurvatureFunction(MRIS *pmris, float (*f)(float k1, float k2))
 {
   //
   // PRECONDITIONS
@@ -10469,7 +12572,7 @@ int MRISusePrincipalCurvatureFunction(MRI_SURFACE *pmris, float (*f)(float k1, f
   Drury and Van Essen:  ICI = 56  (lh) and 54  (rh)
   FI = 500 (lh) and 520 (rh)
   ------------------------------------------------------*/
-int MRIScomputeCurvatureIndices(MRI_SURFACE *mris, double *pici, double *pfi)
+int MRIScomputeCurvatureIndices(MRIS *mris, double *pici, double *pfi)
 {
   int vno;
   VERTEX *vertex;
@@ -10506,7 +12609,7 @@ int MRIScomputeCurvatureIndices(MRI_SURFACE *mris, double *pici, double *pfi)
 
   Description
   ------------------------------------------------------*/
-double MRISmaxRadius(MRI_SURFACE *mris)
+double MRISmaxRadius(MRIS *mris)
 {
   double radius;
   int vno, n;
@@ -10581,7 +12684,7 @@ double MRISmaxRadius(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-double MRISaverageRadius(MRI_SURFACE *mris)
+double MRISaverageRadius(MRIS *mris)
 {
   double radius;
   int vno, n;
@@ -10644,14 +12747,90 @@ double MRISaverageRadius(MRI_SURFACE *mris)
   radius /= (double)n;
   return (radius);
 }
-/*-----------------------------------------------------
-  Parameters:
 
-  Returns value:
 
-  Descripton
-  ------------------------------------------------------*/
-double MRISaverageCanonicalRadius(MRI_SURFACE *mris)
+/* --------------------------------------------------------------------
+   MRISrescaleMetricProperties() - rescale metric properties (area,
+   dist) of group surfaces so that they match that of the average of
+   the input set. Does not change the vertex xyz. Requires that
+   TBD() has been run. Returns the distance scaling factor. 
+   Has no effect if surface is not a group surface.
+   -------------------------------------------------------------------- */
+double MRISrescaleMetricProperties(MRIS *surf)
+{
+  int VtxNo, nthNNbr, nNNbrs, NbrVtxNo;
+  double scale;
+
+  if (surf->group_avg_surface_area == 0) {
+    return (0.0);  // not a group surface
+  }
+
+  scale = sqrt((double)surf->group_avg_surface_area / surf->total_area);
+  printf("scale = %lf\n", scale);
+
+  for (VtxNo = 0; VtxNo < surf->nvertices; VtxNo++) {
+    VERTEX * const vtx1 = &surf->vertices[VtxNo];
+    if (vtx1->ripflag) {
+      continue;
+    }
+    vtx1->area *= (scale * scale);
+    nNNbrs = surf->vertices_topology[VtxNo].vnum;
+    for (nthNNbr = 0; nthNNbr < nNNbrs; nthNNbr++) {
+      NbrVtxNo = surf->vertices_topology[VtxNo].v[nthNNbr];
+      VERTEX const * const vtx2 = &surf->vertices[NbrVtxNo];
+      if (vtx2->ripflag) {
+        continue;
+      }
+      vtx1->dist[nthNNbr] *= scale;
+    }
+  }
+  surf->total_area *= (scale * scale);
+  mrisSetAvgInterVertexDist(surf, surf->avg_vertex_dist * scale);
+  surf->std_vertex_dist *= scale;
+  surf->avg_vertex_area *= (scale * scale);
+  return (scale);
+}
+
+
+int MRISnormalize(MRIS *mris, int dof, int which)
+{
+  int vno;
+  VERTEX *v;
+  float fdof = (float)dof, mean;
+
+  if (dof <= 0) ErrorReturn(ERROR_BADPARM, (ERROR_BADPARM, "MRISnormalize: invalid dof %d", dof));
+
+  for (vno = 0; vno < mris->nvertices; vno++) {
+    v = &mris->vertices[vno];
+    if (v->ripflag) {
+      continue;
+    }
+    if (vno == Gdiag_no) {
+      DiagBreak();
+    }
+    switch (which) {
+      case VERTEX_AREA:
+        v->origarea /= fdof;
+        mean = v->origarea;
+        break;
+      case VERTEX_CURV:
+        v->curv /= fdof;
+        mean = v->curv;
+        break;
+      default:
+      case VERTEX_LOGODDS:
+      case VERTEX_VAL:
+        v->val /= fdof;
+        mean = v->val;
+        break;
+    }
+    v->val2 = v->val2 / fdof - mean * mean;
+  }
+  return (NO_ERROR);
+}
+
+
+double MRISaverageCanonicalRadius(MRIS *mris)
 {
   double radius;
   int vno, n;
@@ -10719,7 +12898,7 @@ double MRISaverageCanonicalRadius(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-int MRISupdateSurface(MRI_SURFACE *mris)
+int MRISupdateSurface(MRIS *mris)
 {
   MRIScomputeMetricProperties(mris);
   MRIScomputeSecondFundamentalForm(mris);
@@ -10817,7 +12996,7 @@ int mriSurfaceRASToVoxel(
 //////////////////////////////////////////////////////////////
 // Find inside and outside voxels of a surface mris
 //////////////////////////////////////////////////////////////
-MRI *MRISbinarizeVolume(MRI_SURFACE *mris, MRI_REGION *region, float resolution, float distance_from_surface)
+MRI *MRISbinarizeVolume(MRIS *mris, MRI_REGION *region, float resolution, float distance_from_surface)
 {
   int k, i, j, p, width, height, depth;
   float x0, x1, x2, y0, y1, y2, z0, z1, z2;
@@ -11110,7 +13289,7 @@ MRI *MRISbinarizeVolume(MRI_SURFACE *mris, MRI_REGION *region, float resolution,
    constituting the tessellation. Writes the orientation changes in
    curv. Note that this function can easily be modified to check the
    orientation changes for any tessellations (not only triangulations) */
-int MRISmarkOrientationChanges(MRI_SURFACE *mris)
+int MRISmarkOrientationChanges(MRIS *mris)
 {
   int face1, face2, vn0, p, vn1, vn2, d1, d2, count;
 
@@ -11215,7 +13394,7 @@ int MRISmarkOrientationChanges(MRI_SURFACE *mris)
   ------------------------------------------------------*/
 #if (!SPHERE_INTERSECTION)
 static int mrisComputeCanonicalEdgeBasis(
-    MRI_SURFACE *mris, EDGE *edge1, EDGE *edge2, double origin[3], double e0[3], double e1[3])
+    MRIS *mris, EDGE *edge1, EDGE *edge2, double origin[3], double e0[3], double e1[3])
 {
   VERTEX *v0, *v1, *v2, *v3;
   double len, normal[3];
@@ -11264,7 +13443,7 @@ static int mrisComputeCanonicalEdgeBasis(
 
   Description
   ------------------------------------------------------*/
-static int mrisDumpDefectiveEdge(MRI_SURFACE *mris, int vno1, int vno2)
+static int mrisDumpDefectiveEdge(MRIS *mris, int vno1, int vno2)
 {
 #if 0
   FILE   *fp ;
@@ -11342,7 +13521,7 @@ static int mrisDumpDefectiveEdge(MRI_SURFACE *mris, int vno1, int vno2)
 
   Description
   ------------------------------------------------------*/
-int mrisCheckSurface(MRI_SURFACE *mris)
+int mrisCheckSurface(MRIS *mris)
 {
   int    fno, vno, n, nfaces, m, vno2, nbad, flist[100];
 
@@ -11447,7 +13626,7 @@ int mrisCheckSurface(MRI_SURFACE *mris)
   ------------------------------------------------------*/
 #if 0
 static int
-mrisDumpTriangle(MRI_SURFACE *mris, int fno)
+mrisDumpTriangle(MRIS *mris, int fno)
 {
   char   fname[STRLEN] ;
   VERTEX *v0, *v1, *v2 ;
@@ -11504,7 +13683,7 @@ mrisDumpTriangle(MRI_SURFACE *mris, int fno)
 
   Description
   ------------------------------------------------------*/
-int MRISmarkNegativeVertices(MRI_SURFACE *mris, int mark)
+int MRISmarkNegativeVertices(MRIS *mris, int mark)
 {
   int fno, n;
   FACE *f;
@@ -11525,7 +13704,7 @@ int MRISmarkNegativeVertices(MRI_SURFACE *mris, int mark)
 
   Description
   ------------------------------------------------------*/
-int MRISripNegativeVertices(MRI_SURFACE *mris)
+int MRISripNegativeVertices(MRIS *mris)
 {
   int fno, n;
   FACE *f;
@@ -11549,7 +13728,7 @@ int MRISripNegativeVertices(MRI_SURFACE *mris)
 
   Description
   ------------------------------------------------------*/
-double MRIScomputeAverageCurvature(MRI_SURFACE *mris, double *psigma)
+double MRIScomputeAverageCurvature(MRIS *mris, double *psigma)
 {
   double mean, var, total, total_sq, nv, d;
   int vno;
@@ -11579,7 +13758,7 @@ double MRIScomputeAverageCurvature(MRI_SURFACE *mris, double *psigma)
 }
 
 
-int MRISmaskLabel(MRI_SURFACE *mris, LABEL *area)
+int MRISmaskLabel(MRIS *mris, LABEL *area)
 {
   int i;
   VERTEX *v;
@@ -11591,7 +13770,7 @@ int MRISmaskLabel(MRI_SURFACE *mris, LABEL *area)
   return (NO_ERROR);
 }
 
-int MRISmaskNotLabel(MRI_SURFACE *mris, LABEL *area)
+int MRISmaskNotLabel(MRIS *mris, LABEL *area)
 {
   int i, vno;
   VERTEX *v;
@@ -11614,7 +13793,7 @@ int MRISmaskNotLabel(MRI_SURFACE *mris, LABEL *area)
   return (NO_ERROR);
 }
 
-int MRISsubsampleDist(MRI_SURFACE *mris, float spacing)
+int MRISsubsampleDist(MRIS *mris, float spacing)
 {
   int k, m, n, sub_num;
 
@@ -12070,7 +14249,7 @@ int **MRIScrsLUT(MRIS *surf, MRI *src)
   return (crslut);
 }
 /*----------------------------------------------------------
-  MRIScrsLUTFree() - frees memoary allocated by MRIScrsLUT.
+  MRIScrsLUTFree() - frees memory allocated by MRIScrsLUT.
   ----------------------------------------------------------*/
 int MRIScrsLUTFree(int **crslut)
 {
@@ -12079,213 +14258,6 @@ int MRIScrsLUTFree(int **crslut)
   free(crslut[2]);
   free(crslut);
   return (0);
-}
-
-int MRISmarkedSpringTerm(MRI_SURFACE *mris, double l_spring)
-{
-  int vno, n, m;
-  float sx, sy, sz, x, y, z;
-
-  if (FZERO(l_spring)) {
-    return (NO_ERROR);
-  }
-
-  for (vno = 0; vno < mris->nvertices; vno++) {
-    VERTEX_TOPOLOGY const * const vertext = &mris->vertices_topology[vno];
-    VERTEX                * const vertex  = &mris->vertices         [vno];
-
-    if (vertex->ripflag || vertex->marked == 0) {
-      continue;
-    }
-    if (vno == Gdiag_no) {
-      DiagBreak();
-    }
-
-    x = vertex->x;
-    y = vertex->y;
-    z = vertex->z;
-
-    sx = sy = sz = 0.0;
-
-    n = 0;
-    for (m = 0; m < vertext->vnum; m++) {
-      VERTEX const * const vn = &mris->vertices[vertext->v[m]];
-      if (!vn->ripflag) {
-        sx += vn->x - x;
-        sy += vn->y - y;
-        sz += vn->z - z;
-        n++;
-      }
-    }
-    if (n > 0) {
-      sx = sx / n;
-      sy = sy / n;
-      sz = sz / n;
-    }
-    sx = l_spring * sx; /* move in normal direction */
-    sy = l_spring * sy;
-    sz = l_spring * sz;
-
-    vertex->dx += sx;
-    vertex->dy += sy;
-    vertex->dz += sz;
-    if (vno == Gdiag_no) fprintf(stdout, "v %d marked spring term:  (%2.3f, %2.3f, %2.3f)\n", vno, sx, sy, sz);
-  }
-
-  return (NO_ERROR);
-}
-
-int MRISnormalSpringTermWithGaussianCurvature(MRI_SURFACE *mris, double gaussian_norm, double l_spring)
-{
-  int vno, n, m;
-  float sx, sy, sz, x, y, z, scale, nc, nx, ny, nz;
-
-  if (FZERO(l_spring)) {
-    return (NO_ERROR);
-  }
-
-  for (vno = 0; vno < mris->nvertices; vno++) {
-    VERTEX_TOPOLOGY const * const vertext = &mris->vertices_topology[vno];
-    VERTEX                * const vertex  = &mris->vertices         [vno];
-    if (vertex->ripflag) {
-      continue;
-    }
-    if (vno == Gdiag_no) {
-      DiagBreak();
-    }
-
-    x = vertex->x;
-    y = vertex->y;
-    z = vertex->z;
-    nx = vertex->nx;
-    ny = vertex->ny;
-    nz = vertex->nz;
-
-    sx = sy = sz = 0.0;
-
-    n = 0;
-    for (m = 0; m < vertext->vnum; m++) {
-      VERTEX const * const vn = &mris->vertices[vertext->v[m]];
-      if (!vn->ripflag) {
-        sx += vn->x - x;
-        sy += vn->y - y;
-        sz += vn->z - z;
-        n++;
-      }
-    }
-    if (n > 0) {
-      sx = sx / n;
-      sy = sy / n;
-      sz = sz / n;
-    }
-    nc = sx * nx + sy * ny + sz * nz; /* projection onto normal */
-    sx = l_spring * nc * nx;          /* move in normal direction */
-    sy = l_spring * nc * ny;
-    sz = l_spring * nc * nz;
-    scale = pow(fabs(vertex->K), gaussian_norm);
-    if (!isfinite(scale)) {
-      scale = 0;
-    };
-    if (scale > 1) {
-      scale = 1;
-    }
-    scale *= l_spring;
-    sx *= scale; /* move in normal direction */
-    sy *= scale;
-    sz *= scale;
-
-    vertex->dx += sx;
-    vertex->dy += sy;
-    vertex->dz += sz;
-    if (vno == Gdiag_no) fprintf(stdout, "v %d Gaussian normal term:  (%2.3f, %2.3f, %2.3f)\n", vno, sx, sy, sz);
-  }
-
-  return (NO_ERROR);
-}
-int MRISspringTermWithGaussianCurvature(MRI_SURFACE *mris, double gaussian_norm, double l_spring)
-{
-  int vno, n, m;
-  float sx, sy, sz, x, y, z, scale;
-
-  if (FZERO(l_spring)) {
-    return (NO_ERROR);
-  }
-
-  for (vno = 0; vno < mris->nvertices; vno++) {
-    VERTEX_TOPOLOGY const * const vertext = &mris->vertices_topology[vno];
-    VERTEX                * const vertex  = &mris->vertices         [vno];
-    if (vertex->ripflag) {
-      continue;
-    }
-    if (vno == Gdiag_no) {
-      DiagBreak();
-    }
-
-    x = vertex->x;
-    y = vertex->y;
-    z = vertex->z;
-
-    sx = sy = sz = 0.0;
-    n = 0;
-    for (m = 0; m < vertext->vnum; m++) {
-      VERTEX const * const vn = &mris->vertices[vertext->v[m]];
-      if (!vn->ripflag) {
-        sx += vn->x - x;
-        sy += vn->y - y;
-        sz += vn->z - z;
-        n++;
-      }
-    }
-    if (n > 0) {
-      sx = sx / n;
-      sy = sy / n;
-      sz = sz / n;
-    }
-    scale = pow(fabs(vertex->K), gaussian_norm);
-    if (!isfinite(scale)) {
-      scale = 0;
-    };
-    if (scale > 1) {
-      scale = 1;
-    }
-    scale *= l_spring;
-    sx *= scale; /* move in normal direction */
-    sy *= scale;
-    sz *= scale;
-
-    vertex->dx += sx;
-    vertex->dy += sy;
-    vertex->dz += sz;
-    if (vno == Gdiag_no) fprintf(stdout, "v %d Gaussian normal term:  (%2.3f, %2.3f, %2.3f)\n", vno, sx, sy, sz);
-  }
-
-  return (NO_ERROR);
-}
-
-int MRISnormalTermWithGaussianCurvature(MRI_SURFACE *mris, double lambda)
-{
-  int vno;
-  VERTEX *v;
-
-  if (FZERO(lambda)) {
-    return (NO_ERROR);
-  }
-
-  for (vno = 0; vno < mris->nvertices; vno++) {
-    v = &mris->vertices[vno];
-    if (v->ripflag) {
-      continue;
-    }
-    if (vno == Gdiag_no) {
-      DiagBreak();
-    }
-
-    v->dx = -v->nx * v->K * lambda;
-    v->dy = -v->ny * v->K * lambda;
-    v->dz = -v->nz * v->K * lambda;
-  }
-
-  return (NO_ERROR);
 }
 
 /*-------------------------------------------------------------------
@@ -12418,7 +14390,7 @@ int MRISextendedNeighbors(MRIS *SphSurf,
 }
 
 
-int mrisMarkIntersections(MRI_SURFACE *mris)
+int mrisMarkIntersections(MRIS *mris)
 {
   MRIS_HASH_TABLE *mht;
   FACE *f;
@@ -12446,7 +14418,7 @@ int mrisMarkIntersections(MRI_SURFACE *mris)
 #define GM_VAL 2
 #define CSF_VAL 3
 
-MRI *MRIcomputeLaminarVolumeFractions(MRI_SURFACE *mris, double resolution, MRI *mri_src, MRI *mri_fractions)
+MRI *MRIcomputeLaminarVolumeFractions(MRIS *mris, double resolution, MRI *mri_src, MRI *mri_fractions)
 {
   int width, height, depth, nvox;
   MRI *mri_layers, *mri_interior_pial, *mri_tmp;
@@ -12537,4 +14509,5 @@ MRI *MRIcomputeLaminarVolumeFractions(MRI_SURFACE *mris, double resolution, MRI 
   MRIScomputeMetricProperties(mris);
   return (mri_fractions);
 }
+
